chore: move methods from utils to network and update coins info
