+++ packages/sysweb3-network/src/index.ts
@@ -2,3 +2,4 @@ export * from './providers/web3';
 export * from './fetch-with-timeout';
 export * from './rpc';
 export * from './rpc-request';
+export * from './networks';
