+++ packages/sysweb3-utils/package.json
@@ -1,6 +1,6 @@
 {
   "name": "@pollum-io/sysweb3-utils",
-  "version": "1.1.195",
+  "version": "1.1.196",
   "description": "A helper for multi-chain accounts.",
   "main": "cjs/index.js",
   "types": "types/index.d.ts",
@@ -25,7 +25,7 @@
   "dependencies": {
     "@ethersproject/contracts": "^5.6.2",
     "@pollum-io/sysweb3-core": "^1.0.16",
-    "@pollum-io/sysweb3-network": "^1.0.55",
+    "@pollum-io/sysweb3-network": "^1.0.59",
     "axios": "^0.26.1",
     "bech32": "^2.0.0",
     "bip32": "^3.0.1",
