--- packages/sysweb3-utils/src/data.ts
@@ -1,1145 +1,0 @@
-import { ICoinCreated } from 'coinsCreation';
-
-export const coinsWithDefaultWif: ICoinCreated[] = [
-  {
-    name: 'Aquariuscoin',
-    networkVersion: 0x17,
-    wif: 0x97,
-    wifStart: '6',
-    cWifStart: 'P',
-  },
-  {
-    name: 'Axe',
-    networkVersion: 0x4b,
-    wif: 0xcb,
-    wifStart: '7',
-    cWifStart: 'X',
-  },
-  {
-    name: 'BBQcoin',
-    networkVersion: 0x55,
-    wif: 0xd5,
-    wifStart: '6',
-    cWifStart: 'T',
-  },
-  {
-    name: 'Biblepay',
-    networkVersion: 0x19,
-    wif: 0xb6,
-    wifStart: '7',
-    cWifStart: '[TU]',
-  },
-  {
-    name: 'Bitcoin',
-    networkVersion: 0x00,
-    wif: 0x80,
-    wifStart: '5',
-    cWifStart: '[LK]',
-  },
-  {
-    name: 'BitcoinCash',
-    networkVersion: 0x00,
-    wif: 0x80,
-    wifStart: '5',
-    cWifStart: '[LK]',
-  },
-  {
-    name: 'BitcoinDark',
-    networkVersion: 0x3c,
-    wif: 0xbc,
-    wifStart: '7',
-    cWifStart: 'U',
-  },
-  {
-    name: 'Bitcore',
-    networkVersion: 0x00,
-    wif: 0x80,
-    wifStart: '5',
-    cWifStart: '[LK]',
-  },
-  {
-    name: 'BitcoinGold',
-    networkVersion: 0x26,
-    wif: 0x80,
-    wifStart: '5',
-    cWifStart: '[LK]',
-  },
-  {
-    name: 'Bitconnect',
-    networkVersion: 0x12,
-    wif: 0x92,
-    wifStart: '5',
-    cWifStart: 'N',
-  },
-  {
-    name: 'Birdcoin',
-    networkVersion: 0x2f,
-    wif: 0xaf,
-    wifStart: '6',
-    cWifStart: '[ST]',
-  },
-  {
-    name: 'BitSynq',
-    networkVersion: 0x3f,
-    wif: 0xbf,
-    wifStart: '7',
-    cWifStart: 'V',
-  },
-  {
-    name: 'BitZeny',
-    networkVersion: 0x51,
-    wif: 0x80,
-    wifStart: '5',
-    cWifStart: '[LK]',
-  },
-  {
-    name: 'Blackcoin',
-    networkVersion: 0x19,
-    wif: 0x99,
-    wifStart: '6',
-    cWifStart: 'P',
-  },
-  {
-    name: 'BlackJack',
-    networkVersion: 0x15,
-    wif: 0x95,
-    wifStart: '[56]',
-    cWifStart: 'P',
-  },
-  {
-    name: 'BlockNet',
-    networkVersion: 0x1a,
-    wif: 0x9a,
-    wifStart: '6',
-    cWifStart: 'P',
-  },
-  {
-    name: 'BolivarCoin',
-    networkVersion: 0x55,
-    wif: 0xd5,
-    wifStart: '8',
-    cWifStart: 'Y',
-  },
-  {
-    name: 'BoxyCoin',
-    networkVersion: 0x4b,
-    wif: 0xcb,
-    wifStart: '7',
-    cWifStart: 'X',
-  },
-  {
-    name: 'BunnyCoin',
-    networkVersion: 0x1a,
-    wif: 0x9a,
-    wifStart: '6',
-    cWifStart: 'P',
-  },
-  {
-    name: 'Cagecoin',
-    networkVersion: 0x1f,
-    wif: 0x9f,
-    wifStart: '6',
-    cWifStart: 'Q',
-  },
-  {
-    name: 'CampusCoin',
-    networkVersion: 0x1c,
-    wif: 0x9c,
-    wifStart: '6',
-    cWifStart: 'Q',
-  },
-  {
-    name: 'CanadaeCoin',
-    networkVersion: 0x1c,
-    wif: 0x9c,
-    wifStart: '6',
-    cWifStart: 'Q',
-  },
-  {
-    name: 'CannabisCoin',
-    networkVersion: 0x1c,
-    wif: 0x9c,
-    wifStart: '6',
-    cWifStart: 'Q',
-  },
-  {
-    name: 'Capricoin',
-    networkVersion: 0x1c,
-    wif: 0x9c,
-    wifStart: '6',
-    cWifStart: 'Q',
-  },
-  {
-    name: 'CassubianDetk',
-    networkVersion: 0x1e,
-    wif: 0x9e,
-    wifStart: '6',
-    cWifStart: 'Q',
-  },
-  {
-    name: 'CashCoin',
-    networkVersion: 0x22,
-    wif: 0xa2,
-    wifStart: '6',
-    cWifStart: '[QR]',
-  },
-  {
-    name: 'Catcoin',
-    networkVersion: 0x15,
-    wif: 0x95,
-    wifStart: '[56]',
-    cWifStart: 'P',
-  },
-  {
-    name: 'ChainCoin',
-    networkVersion: 0x1c,
-    wif: 0x9c,
-    wifStart: '6',
-    cWifStart: 'Q',
-  },
-  {
-    name: 'ColossusCoinXT',
-    networkVersion: 0x1e,
-    wif: 0xd4,
-    wifStart: '5',
-    cWifStart: '[LK]',
-  },
-  {
-    name: 'Condensate',
-    networkVersion: 0x3c,
-    wif: 0xbc,
-    wifStart: '7',
-    cWifStart: 'U',
-  },
-  {
-    name: 'Copico',
-    networkVersion: 0x1c,
-    wif: 0x90,
-    wifStart: '5',
-    cWifStart: 'N',
-  },
-  {
-    name: 'CopperCoin',
-    networkVersion: 0x1c,
-    wif: 0x9c,
-    wifStart: '6',
-    cWifStart: 'Q',
-  },
-  {
-    name: 'Corgicoin',
-    networkVersion: 0x1c,
-    wif: 0x9c,
-    wifStart: '6',
-    cWifStart: 'Q',
-  },
-  {
-    name: 'CryptoBullion',
-    networkVersion: 0x0b,
-    wif: 0x8b,
-    wifStart: '5',
-    cWifStart: 'M',
-  },
-  {
-    name: 'CryptoClub',
-    networkVersion: 0x23,
-    wif: 0xa3,
-    wifStart: '6',
-    cWifStart: 'R',
-  },
-  {
-    name: 'Cryptoescudo',
-    networkVersion: 0x1c,
-    wif: 0x9c,
-    wifStart: '6',
-    cWifStart: 'Q',
-  },
-  {
-    name: 'Cryptonite',
-    networkVersion: 0x1c,
-    wif: 0x80,
-    wifStart: '5',
-    cWifStart: '[LK]',
-  },
-  {
-    name: 'CryptoWisdomCoin',
-    networkVersion: 0x49,
-    wif: 0x87,
-    wifStart: '5',
-    cWifStart: '[LM]',
-  },
-  {
-    name: 'C2coin',
-    networkVersion: 0x1c,
-    wif: 0x9c,
-    wifStart: '6',
-    cWifStart: 'Q',
-  },
-  {
-    name: 'Dash',
-    networkVersion: 0x4c,
-    wif: 0xcc,
-    wifStart: '7',
-    cWifStart: 'X',
-  },
-  {
-    name: 'DeafDollars',
-    networkVersion: 0x30,
-    wif: 0xb0,
-    wifStart: '6',
-    cWifStart: 'T',
-  },
-  {
-    name: 'DeepOnion',
-    networkVersion: 0x1f,
-    wif: 0x9f,
-    wifStart: '6',
-    cWifStart: 'Q',
-  },
-  {
-    name: 'Deutsche eMark',
-    networkVersion: 0x35,
-    wif: 0xb5,
-    wifStart: '7',
-    cWifStart: 'T',
-  },
-  {
-    name: 'Devcoin',
-    networkVersion: 0x00,
-    wif: 0x80,
-    wifStart: '5',
-    cWifStart: '[LK]',
-  },
-  {
-    name: 'DigiByte',
-    networkVersion: 0x1e,
-    wif: 0x9e,
-    wifStart: '6',
-    cWifStart: 'Q',
-  },
-  {
-    name: 'Digitalcoin',
-    networkVersion: 0x1e,
-    wif: 0x9e,
-    wifStart: '6',
-    cWifStart: 'Q',
-  },
-  {
-    name: 'Dimecoin',
-    networkVersion: 0x0f,
-    wif: 0x8f,
-    wifStart: '5',
-    cWifStart: 'N',
-  },
-  {
-    name: 'DNotes',
-    networkVersion: 0x1f,
-    wif: 0x9f,
-    wifStart: '6',
-    cWifStart: 'Q',
-  },
-  {
-    name: 'Dogecoin',
-    networkVersion: 0x1e,
-    wif: 0x9e,
-    wifStart: '6',
-    cWifStart: 'Q',
-  },
-  {
-    name: 'DogecoinDark',
-    networkVersion: 0x1e,
-    wif: 0x9e,
-    wifStart: '6',
-    cWifStart: 'Q',
-  },
-  {
-    name: 'eGulden',
-    networkVersion: 0x30,
-    wif: 0xb0,
-    wifStart: '6',
-    cWifStart: 'T',
-  },
-  {
-    name: 'eKrona',
-    networkVersion: 0x2d,
-    wif: 0xad,
-    wifStart: '6',
-    cWifStart: 'S',
-  },
-  {
-    name: 'ELECTRA',
-    networkVersion: 0x21,
-    wif: 0xa1,
-    wifStart: '6',
-    cWifStart: 'Q',
-  },
-  {
-    name: 'Ember',
-    networkVersion: 0x5c,
-    wif: 0x32,
-    wifStart: '2',
-    cWifStart: '8',
-  },
-  {
-    name: 'Emerald',
-    networkVersion: 0x22,
-    wif: 0xa2,
-    wifStart: '6',
-    cWifStart: '[QR]',
-  },
-  {
-    name: 'Emercoin',
-    networkVersion: 0x21,
-    wif: 0x80,
-    wifStart: '5',
-    cWifStart: '[LK]',
-  },
-  {
-    name: 'EnergyCoin',
-    networkVersion: 0x5c,
-    wif: 0xdc,
-    wifStart: '8',
-    cWifStart: 'Z',
-  },
-  {
-    name: 'Espers',
-    networkVersion: 0x21,
-    wif: 0x90,
-    wifStart: '5',
-    cWifStart: 'N',
-  },
-  {
-    name: 'Fastcoin',
-    networkVersion: 0x60,
-    wif: 0xe0,
-    wifStart: '8',
-    cWifStart: 'a',
-  },
-  {
-    name: 'Feathercoin',
-    networkVersion: 0x0e,
-    wif: 0x8e,
-    wifStart: '5',
-    cWifStart: 'N',
-  },
-  {
-    name: 'Fedoracoin',
-    networkVersion: 0x21,
-    wif: 0x80,
-    wifStart: '5',
-    cWifStart: '[KL]',
-  },
-  {
-    name: 'Fibre',
-    networkVersion: 0x23,
-    wif: 0xa3,
-    wifStart: '6',
-    cWifStart: 'R',
-  },
-  {
-    name: 'Florincoin',
-    networkVersion: 0x23,
-    wif: 0xb0,
-    wifStart: '6',
-    cWifStart: 'T',
-  },
-  {
-    name: 'Flurbo',
-    networkVersion: 0x23,
-    wif: 0x30,
-    wifStart: '6',
-    cWifStart: '8',
-  },
-  {
-    name: 'Fluttercoin',
-    networkVersion: 0x23,
-    wif: 0xa3,
-    wifStart: '6',
-    cWifStart: 'R',
-  },
-  {
-    name: 'FrazCoin',
-    networkVersion: 0x23,
-    wif: 0xa3,
-    wifStart: '6',
-    cWifStart: 'R',
-  },
-  {
-    name: 'Freicoin',
-    networkVersion: 0x00,
-    wif: 0x80,
-    wifStart: '5',
-    cWifStart: '[LK]',
-  },
-  {
-    name: 'FUDcoin',
-    networkVersion: 0x23,
-    wif: 0xa3,
-    wifStart: '6',
-    cWifStart: 'R',
-  },
-  {
-    name: 'Fuelcoin',
-    networkVersion: 0x24,
-    wif: 0x80,
-    wifStart: '5',
-    cWifStart: '[KL]',
-  },
-  {
-    name: 'Fujicoin',
-    networkVersion: 0x24,
-    wif: 0xa4,
-    wifStart: '6',
-    cWifStart: 'R',
-  },
-  {
-    name: 'GabenCoin',
-    networkVersion: 0x10,
-    wif: 0x90,
-    wifStart: '5',
-    cWifStart: 'N',
-  },
-  {
-    name: 'Garlicoin',
-    networkVersion: 0x26,
-    wif: 0xb0,
-    wifStart: '6',
-    cWifStart: 'T',
-  },
-  {
-    name: 'GlobalBoost',
-    networkVersion: 0x26,
-    wif: 0xa6,
-    wifStart: '6',
-    cWifStart: 'R',
-  },
-  {
-    name: 'Goodcoin',
-    networkVersion: 0x26,
-    wif: 0xa6,
-    wifStart: '6',
-    cWifStart: 'R',
-  },
-  {
-    name: 'GridcoinResearch',
-    networkVersion: 0x3e,
-    wif: 0xbe,
-    wifStart: '7',
-    cWifStart: 'V',
-  },
-  {
-    name: 'Gulden',
-    networkVersion: 0x26,
-    wif: 0xa6,
-    wifStart: '6',
-    cWifStart: 'R',
-  },
-  {
-    name: 'Guncoin',
-    networkVersion: 0x27,
-    wif: 0xa7,
-    wifStart: '6',
-    cWifStart: 'R',
-  },
-  {
-    name: 'HamRadioCoin',
-    networkVersion: 0x00,
-    wif: 0x80,
-    wifStart: '5',
-    cWifStart: '[LK]',
-  },
-  {
-    name: 'HFRcoin',
-    networkVersion: 0x10,
-    wif: 0x90,
-    wifStart: '5',
-    cWifStart: 'N',
-  },
-  {
-    name: 'HOdlcoin',
-    networkVersion: 0x28,
-    wif: 0xa8,
-    wifStart: '5',
-    cWifStart: 'LK',
-  },
-  {
-    name: 'HTMLCoin',
-    networkVersion: 0x29,
-    wif: 0xa9,
-    wifStart: '6',
-    cWifStart: 'S',
-  },
-  {
-    name: 'HyperStake',
-    networkVersion: 0x75,
-    wif: 0xf5,
-    wifStart: '9',
-    cWifStart: 'd',
-  },
-  {
-    name: 'ImperiumCoin',
-    networkVersion: 0x30,
-    wif: 0xb0,
-    wifStart: '6',
-    cWifStart: 'T',
-  },
-  {
-    name: 'IncaKoin',
-    networkVersion: 0x35,
-    wif: 0xb5,
-    wifStart: '7',
-    cWifStart: 'T',
-  },
-  {
-    name: 'IncognitoCoin',
-    networkVersion: 0x00,
-    wif: 0x80,
-    wifStart: '5',
-    cWifStart: 'LK',
-  },
-  {
-    name: 'Influxcoin',
-    networkVersion: 0x66,
-    wif: 0xe6,
-    wifStart: '8',
-    cWifStart: 'b',
-  },
-  {
-    name: 'Innox',
-    networkVersion: 0x4b,
-    wif: 0xcb,
-    wifStart: '7',
-    cWifStart: 'X',
-  },
-  {
-    name: 'IridiumCoin',
-    networkVersion: 0x30,
-    wif: 0xb0,
-    wifStart: '6',
-    cWifStart: 'T',
-  },
-  {
-    name: 'iCash',
-    networkVersion: 0x66,
-    wif: 0xcc,
-    wifStart: '7',
-    cWifStart: 'X',
-  },
-  {
-    name: 'IXcoin',
-    networkVersion: 0x8a,
-    wif: 0x80,
-    wifStart: '5',
-    cWifStart: '[LK]',
-  },
-  {
-    name: 'Judgecoin',
-    networkVersion: 0x2b,
-    wif: 0xab,
-    wifStart: '6',
-    cWifStart: 'S',
-  },
-  {
-    name: 'Jumbucks',
-    networkVersion: 0x2b,
-    wif: 0xab,
-    wifStart: '6',
-    cWifStart: 'S',
-  },
-  {
-    name: 'KHcoin',
-    networkVersion: 0x30,
-    wif: 0xb0,
-    wifStart: '6',
-    cWifStart: 'T',
-  },
-  {
-    name: 'KittehCoin',
-    networkVersion: 0x2d,
-    wif: 0xad,
-    wifStart: '6',
-    cWifStart: 'S',
-  },
-  {
-    name: 'Lanacoin',
-    networkVersion: 0x30,
-    wif: 0xb0,
-    wifStart: '6',
-    cWifStart: 'T',
-  },
-  {
-    name: 'Latium',
-    networkVersion: 0x17,
-    wif: 0x80,
-    wifStart: '5',
-    cWifStart: '[LK]',
-  },
-  {
-    name: 'LBRY Credits',
-    networkVersion: 0x55,
-    wif: 0x80,
-    wifStart: '5',
-    cWifStart: '[LK]',
-  },
-  {
-    name: 'Litecoin',
-    networkVersion: 0x30,
-    wif: 0xb0,
-    wifStart: '6',
-    cWifStart: 'T',
-  },
-  {
-    name: 'LiteDoge',
-    networkVersion: 0x5a,
-    wif: 0xab,
-    wifStart: '6',
-    cWifStart: 'S',
-  },
-  {
-    name: 'LoMoCoin',
-    networkVersion: 0x30,
-    wif: 0xb0,
-    wifStart: '6',
-    cWifStart: 'T',
-  },
-  {
-    name: 'MadbyteCoin',
-    networkVersion: 0x32,
-    wif: 0x6e,
-    wifStart: '4',
-    cWifStart: 'H',
-  },
-  {
-    name: 'MagicInternetMoney',
-    networkVersion: 0x30,
-    wif: 0xb0,
-    wifStart: '6',
-    cWifStart: 'T',
-  },
-  {
-    name: 'Magicoin',
-    networkVersion: 0x14,
-    wif: 0x94,
-    wifStart: '5',
-    cWifStart: '[NP]',
-  },
-  {
-    name: 'Marscoin',
-    networkVersion: 0x32,
-    wif: 0xb2,
-    wifStart: '6',
-    cWifStart: 'T',
-  },
-  {
-    name: 'MarteXcoin',
-    networkVersion: 0x32,
-    wif: 0xb2,
-    wifStart: '6',
-    cWifStart: 'T',
-  },
-  {
-    name: 'MasterDoge',
-    networkVersion: 0x33,
-    wif: 0x8b,
-    wifStart: '5',
-    cWifStart: 'M',
-  },
-  {
-    name: 'Mazacoin',
-    networkVersion: 0x32,
-    wif: 0xe0,
-    wifStart: '8',
-    cWifStart: 'a',
-  },
-  {
-    name: 'Megacoin',
-    networkVersion: 0x32,
-    wif: 0xb2,
-    wifStart: '6',
-    cWifStart: 'T',
-  },
-  {
-    name: 'MintCoin',
-    networkVersion: 0x33,
-    wif: 0xb3,
-    wifStart: '[67]',
-    cWifStart: 'T',
-  },
-  {
-    name: 'MobiusCoin',
-    networkVersion: 0x00,
-    wif: 0x80,
-    wifStart: '5',
-    cWifStart: '[LK]',
-  },
-  {
-    name: 'MonetaryUnit',
-    networkVersion: 0x10,
-    wif: 0x7e,
-    wifStart: '5',
-    cWifStart: 'K',
-  },
-  {
-    name: 'Monocle',
-    networkVersion: 0x32,
-    wif: 0xb2,
-    wifStart: '6',
-    cWifStart: 'T',
-  },
-  {
-    name: 'MoonCoin',
-    networkVersion: 0x03,
-    wif: 0x83,
-    wifStart: '5',
-    cWifStart: 'L',
-  },
-  {
-    name: 'Myriadcoin',
-    networkVersion: 0x32,
-    wif: 0xb2,
-    wifStart: '6',
-    cWifStart: 'T',
-  },
-  {
-    name: 'NameCoin',
-    networkVersion: 0x34,
-    wif: 0x80,
-    wifStart: '5',
-    cWifStart: '[LK]',
-  },
-  {
-    name: 'Navcoin',
-    networkVersion: 0x35,
-    wif: 0x96,
-    wifStart: '6',
-    cWifStart: 'P',
-  },
-  {
-    name: 'NeedleCoin',
-    networkVersion: 0x35,
-    wif: 0xb5,
-    wifStart: '7',
-    cWifStart: 'T',
-  },
-  {
-    name: 'NEETCOIN',
-    networkVersion: 0x35,
-    wif: 0xb5,
-    wifStart: '7',
-    cWifStart: 'T',
-  },
-  {
-    name: 'NYC',
-    networkVersion: 0x3c,
-    wif: 0xbc,
-    wifStart: '7',
-    cWifStart: 'U',
-  },
-  {
-    name: 'Neoscoin',
-    networkVersion: 0x35,
-    wif: 0xb1,
-    wifStart: '6',
-    cWifStart: 'T',
-  },
-  {
-    name: 'Nevacoin',
-    networkVersion: 0x35,
-    wif: 0xb1,
-    wifStart: '6',
-    cWifStart: 'T',
-  },
-  {
-    name: 'Novacoin',
-    networkVersion: 0x08,
-    wif: 0x88,
-    wifStart: '5',
-    cWifStart: 'M',
-  },
-  {
-    name: 'Nubits',
-    networkVersion: 0x19,
-    wif: 0xbf,
-    wifStart: '7',
-    cWifStart: 'V',
-  },
-  {
-    name: 'Nyancoin',
-    networkVersion: 0x2d,
-    wif: 0xad,
-    wifStart: '6',
-    cWifStart: 'S',
-  },
-  {
-    name: 'Ocupy',
-    networkVersion: 0x73,
-    wif: 0xf3,
-    wifStart: '9',
-    cWifStart: '[cd]',
-  },
-  {
-    name: 'Omnicoin',
-    networkVersion: 0x73,
-    wif: 0xf3,
-    wifStart: '9',
-    cWifStart: '[cd]',
-  },
-  {
-    name: 'Onyxcoin',
-    networkVersion: 0x73,
-    wif: 0xf3,
-    wifStart: '9',
-    cWifStart: '[cd]',
-  },
-  {
-    name: 'PacCoin',
-    networkVersion: 0x18,
-    wif: 0x98,
-    wifStart: '6',
-    cWifStart: 'P',
-  },
-  {
-    name: 'Particl',
-    networkVersion: 0x38,
-    wif: 0x6c,
-    wifStart: '4',
-    cWifStart: '[HG]',
-  },
-  {
-    name: 'Paycoin',
-    networkVersion: 0x37,
-    wif: 0xb7,
-    wifStart: '7',
-    cWifStart: 'U',
-  },
-  {
-    name: 'Pandacoin',
-    networkVersion: 0x37,
-    wif: 0xb7,
-    wifStart: '7',
-    cWifStart: 'U',
-  },
-  {
-    name: 'ParkByte',
-    networkVersion: 0x37,
-    wif: 0xb7,
-    wifStart: '7',
-    cWifStart: 'U',
-  },
-  {
-    name: 'Peercoin',
-    networkVersion: 0x37,
-    wif: 0xb7,
-    wifStart: '7',
-    cWifStart: 'U',
-  },
-  {
-    name: 'Pesetacoin',
-    networkVersion: 0x2f,
-    wif: 0xaf,
-    wifStart: '6',
-    cWifStart: '[ST]',
-  },
-  {
-    name: 'PHCoin',
-    networkVersion: 0x37,
-    wif: 0xb7,
-    wifStart: '7',
-    cWifStart: 'U',
-  },
-  {
-    name: 'PhoenixCoin',
-    networkVersion: 0x38,
-    wif: 0xb8,
-    wifStart: '7',
-    cWifStart: 'U',
-  },
-  {
-    name: 'PiggyCoin',
-    networkVersion: 0x76,
-    wif: 0xf6,
-    wifStart: '9',
-    cWifStart: 'd',
-  },
-  {
-    name: 'Pinkcoin',
-    networkVersion: 0x3,
-    wif: 0x83,
-    wifStart: '[RQP]',
-    cWifStart: 'L',
-  },
-  {
-    name: 'PIVX',
-    networkVersion: 0x1e,
-    wif: 0xd4,
-    wifStart: '8',
-    cWifStart: 'Y',
-  },
-  {
-    name: 'Peercoin',
-    networkVersion: 0x37,
-    wif: 0xb7,
-    wifStart: '7',
-    cWifStart: 'U',
-  },
-  {
-    name: 'Potcoin',
-    networkVersion: 0x37,
-    wif: 0xb7,
-    wifStart: '7',
-    cWifStart: 'U',
-  },
-  {
-    name: 'Primecoin',
-    networkVersion: 0x17,
-    wif: 0x97,
-    wifStart: '6',
-    cWifStart: 'P',
-  },
-  {
-    name: 'ProsperCoinClassic',
-    networkVersion: 0x3a,
-    wif: 0xba,
-    wifStart: '7',
-    cWifStart: 'Q',
-  },
-  {
-    name: 'Reddcoin',
-    networkVersion: 0x3d,
-    wif: 0xbd,
-    wifStart: '7',
-    cWifStart: '[UV]',
-  },
-  {
-    name: 'Riecoin',
-    networkVersion: 0x3c,
-    wif: 0x80,
-    wifStart: '5',
-    cWifStart: '[LK]',
-  },
-  {
-    name: 'ROIcoin',
-    networkVersion: 0x3c,
-    wif: 0x80,
-    wifStart: '5',
-    cWifStart: '[LK]',
-  },
-  {
-    name: 'Rubycoin',
-    networkVersion: 0x3c,
-    wif: 0xbc,
-    wifStart: '7',
-    cWifStart: 'U',
-  },
-  {
-    name: 'Rupaya',
-    networkVersion: 0x3c,
-    wif: 0xbc,
-    wifStart: '7',
-    cWifStart: 'U',
-  },
-  {
-    name: 'SmileyCoin',
-    networkVersion: 0x19,
-    wif: 0x99,
-    wifStart: '6',
-    cWifStart: 'P',
-  },
-  {
-    name: 'StealthCoin',
-    networkVersion: 0x3e,
-    wif: 0xbe,
-    wifStart: '7',
-    cWifStart: 'V',
-  },
-  {
-    name: 'Stratis',
-    networkVersion: 0x3f,
-    wif: 0xbf,
-    wifStart: '7',
-    cWifStart: 'V',
-  },
-  {
-    name: 'Syscoin',
-    networkVersion: 0x00,
-    wif: 0x80,
-    wifStart: '5',
-    cWifStart: '[LK]',
-  },
-  {
-    name: 'Terracoin',
-    networkVersion: 0x00,
-    wif: 0x80,
-    wifStart: '5',
-    cWifStart: '[LK]',
-  },
-  {
-    name: 'TrezarCoin',
-    networkVersion: 0x42,
-    wif: 0xc2,
-    wifStart: '7',
-    cWifStart: 'V',
-  },
-  {
-    name: 'Unobtanium',
-    networkVersion: 0x82,
-    wif: 0xe0,
-    wifStart: '8',
-    cWifStart: 'a',
-  },
-  {
-    name: 'Vcash',
-    networkVersion: 0x47,
-    wif: 0xc7,
-    wifStart: '7',
-    cWifStart: 'W',
-  },
-  {
-    name: 'Vertcoin',
-    networkVersion: 0x47,
-    wif: 0x80,
-    wifStart: '5',
-    cWifStart: '[LK]',
-  },
-  {
-    name: 'Viacoin',
-    networkVersion: 0x47,
-    wif: 0xc7,
-    wifStart: '7',
-    cWifStart: 'W',
-  },
-  {
-    name: 'Zcash',
-    networkVersion: [0x1c, 0xb8],
-    wif: 0x80,
-    wifStart: '5',
-    cWifStart: '[LK]',
-  },
-  {
-    name: 'Testnet Bitcoin',
-    networkVersion: 0x6f,
-    wif: 0xef,
-    wifStart: '9',
-    cWifStart: 'c',
-  },
-  {
-    name: 'Testnet Dogecoin',
-    networkVersion: 0x71,
-    wif: 0xf1,
-    wifStart: '9',
-    cWifStart: 'c',
-  },
-  {
-    name: 'Testnet MonetaryUnit',
-    networkVersion: 0x26,
-    wif: 0x40,
-    wifStart: '3',
-    cWifStart: 'A',
-  },
-  {
-    name: 'Testnet PIVX',
-    networkVersion: 0x8b,
-    wif: 0xef,
-    wifStart: '9',
-    cWifStart: 'c',
-  },
-];
