+++ packages/sysweb3-utils/src/index.ts
@@ -7,6 +7,7 @@ export * from './transactions';
 export * from './trezor';
 export * from './signer';
 export * from './accounts';
+export * from './coins';
 export * from './networks';
 export * from './spsbt';
 export * from './address';
