--- packages/sysweb3-utils/src/trezorSupportedCoins.ts
@@ -1,2830 +1,0 @@
-/* eslint-disable camelcase */
-
-// todo: add Testnet PIVX and pivx
-export const coins = {
-  bitcoin: [
-    {
-      addressType: 0,
-      addressTypeP2sh: 5,
-      bech32Prefix: 'bc',
-      blockchainLink: {
-        type: 'blockbook',
-        url: [
-          'https://btc1.trezor.io',
-          'https://btc2.trezor.io',
-          'https://btc3.trezor.io',
-          'https://btc4.trezor.io',
-          'https://btc5.trezor.io',
-        ],
-      },
-      blocktimeSeconds: 600,
-      cashaddrPrefix: null,
-      coinLabel: 'Bitcoin',
-      coinName: 'Bitcoin',
-      coinShortcut: 'BTC',
-      curveName: 'secp256k1',
-      decimals: 8,
-      decred: false,
-      defaultFeeB: {
-        economy: 70,
-        high: 200,
-        low: 10,
-        normal: 140,
-      },
-      dustLimit: 546,
-      extraData: false,
-      forceBip143: false,
-      forkId: null,
-      hashGenesisBlock:
-        '000000000019d6689c085ae165831e934ff763ae46a2a6c172b3f1b60a8ce26f',
-      maxAddressLength: 34,
-      maxfeeKb: 2000000,
-      minAddressLength: 27,
-      minfeeKb: 1000,
-      name: 'Bitcoin',
-      segwit: true,
-      shortcut: 'BTC',
-      signedMessageHeader: 'Bitcoin Signed Message:\n',
-      slip44: 0,
-      support: {
-        connect: true,
-        suite: true,
-        trezor1: '1.5.2',
-        trezor2: '2.0.5',
-      },
-      taproot: true,
-      timestamp: false,
-      xprvMagic: 76066276,
-      xpubMagic: 76067358,
-      xpubMagicMultisigSegwitNative: 44728019,
-      xpubMagicMultisigSegwitP2sh: 43365439,
-      xpubMagicSegwitNative: 78792518,
-      xpubMagicSegwitP2sh: 77429938,
-      wif: 80,
-    },
-    {
-      addressType: 111,
-      addressTypeP2sh: 196,
-      bech32Prefix: 'bcrt',
-      blockchainLink: null,
-      blocktimeSeconds: 600,
-      cashaddrPrefix: null,
-      coinLabel: 'Regtest',
-      coinName: 'Regtest',
-      coinShortcut: 'REGTEST',
-      curveName: 'secp256k1',
-      decimals: 8,
-      decred: false,
-      defaultFeeB: {
-        normal: 10,
-      },
-      dustLimit: 546,
-      extraData: false,
-      forceBip143: false,
-      forkId: null,
-      hashGenesisBlock:
-        '0f9188f13cb7b2c71f2a335e3a4fc328bf5beb436012afca590b1a11466e2206',
-      maxAddressLength: 34,
-      maxfeeKb: 10000000,
-      minAddressLength: 27,
-      minfeeKb: 1000,
-      name: 'Regtest',
-      segwit: true,
-      shortcut: 'REGTEST',
-      signedMessageHeader: 'Bitcoin Signed Message:\n',
-      slip44: 1,
-      support: {
-        connect: true,
-        suite: true,
-        trezor1: '1.8.2',
-        trezor2: '2.1.1',
-      },
-      taproot: true,
-      timestamp: false,
-      xprvMagic: 70615956,
-      xpubMagic: 70617039,
-      xpubMagicMultisigSegwitNative: 39277699,
-      xpubMagicMultisigSegwitP2sh: 37915119,
-      xpubMagicSegwitNative: 73342198,
-      xpubMagicSegwitP2sh: 71979618,
-      wif: '',
-    },
-    {
-      addressType: 111,
-      addressTypeP2sh: 196,
-      bech32Prefix: 'tb',
-      blockchainLink: {
-        type: 'blockbook',
-        url: ['https://tbtc1.trezor.io', 'https://tbtc2.trezor.io'],
-      },
-      blocktimeSeconds: 600,
-      cashaddrPrefix: null,
-      coinLabel: 'Testnet',
-      coinName: 'Testnet',
-      coinShortcut: 'TEST',
-      curveName: 'secp256k1',
-      decimals: 8,
-      decred: false,
-      defaultFeeB: {
-        normal: 10,
-      },
-      dustLimit: 546,
-      extraData: false,
-      forceBip143: false,
-      forkId: null,
-      hashGenesisBlock:
-        '000000000933ea01ad0ee984209779baaec3ced90fa3f408719526f8d77f4943',
-      maxAddressLength: 34,
-      maxfeeKb: 10000000,
-      minAddressLength: 27,
-      minfeeKb: 1000,
-      name: 'Bitcoin Testnet',
-      segwit: true,
-      shortcut: 'TEST',
-      signedMessageHeader: 'Bitcoin Signed Message:\n',
-      slip44: 1,
-      support: {
-        connect: true,
-        suite: true,
-        trezor1: '1.5.2',
-        trezor2: '2.0.5',
-      },
-      taproot: true,
-      timestamp: false,
-      xprvMagic: 70615956,
-      xpubMagic: 70617039,
-      xpubMagicMultisigSegwitNative: 39277699,
-      xpubMagicMultisigSegwitP2sh: 37915119,
-      xpubMagicSegwitNative: 73342198,
-      xpubMagicSegwitP2sh: 71979618,
-      wif: '',
-    },
-    {
-      addressType: 53,
-      addressTypeP2sh: 55,
-      bech32Prefix: 'acm',
-      blockchainLink: null,
-      blocktimeSeconds: 150,
-      cashaddrPrefix: null,
-      coinLabel: 'Actinium',
-      coinName: 'Actinium',
-      coinShortcut: 'ACM',
-      curveName: 'secp256k1',
-      decimals: 8,
-      decred: false,
-      defaultFeeB: {
-        normal: 1000,
-      },
-      dustLimit: 546,
-      extraData: false,
-      forceBip143: false,
-      forkId: null,
-      hashGenesisBlock:
-        '28d77872e23714562f49a1be792c276623c1bbe3fdcf21b6035cfde78b00b824',
-      maxAddressLength: 34,
-      maxfeeKb: 320000000000,
-      minAddressLength: 27,
-      minfeeKb: 100000,
-      name: 'Actinium',
-      segwit: true,
-      shortcut: 'ACM',
-      signedMessageHeader: 'Actinium Signed Message:\n',
-      slip44: 228,
-      support: {
-        connect: true,
-        suite: false,
-        trezor1: '1.7.2',
-        trezor2: '2.0.10',
-      },
-      taproot: false,
-      timestamp: false,
-      xprvMagic: 76066276,
-      xpubMagic: 76067358,
-      xpubMagicMultisigSegwitNative: 76067358,
-      xpubMagicMultisigSegwitP2sh: 76067358,
-      xpubMagicSegwitNative: 78792518,
-      xpubMagicSegwitP2sh: 77429938,
-      wif: '',
-    },
-    {
-      addressType: 55,
-      addressTypeP2sh: 16,
-      bech32Prefix: null,
-      blockchainLink: null,
-      blocktimeSeconds: 150,
-      cashaddrPrefix: null,
-      coinLabel: 'Axe',
-      coinName: 'Axe',
-      coinShortcut: 'AXE',
-      curveName: 'secp256k1',
-      decimals: 8,
-      decred: false,
-      defaultFeeB: {
-        normal: 10,
-      },
-      dustLimit: 5460,
-      extraData: false,
-      forceBip143: false,
-      forkId: null,
-      hashGenesisBlock:
-        '00000c33631ca6f2f61368991ce2dc03306b5bb50bf7cede5cfbba6db38e52e6',
-      maxAddressLength: 34,
-      maxfeeKb: 21000000000,
-      minAddressLength: 27,
-      minfeeKb: 1000,
-      name: 'Axe',
-      segwit: false,
-      shortcut: 'AXE',
-      signedMessageHeader: 'DarkCoin Signed Message:\n',
-      slip44: 4242,
-      support: {
-        connect: true,
-        suite: false,
-        trezor1: '1.7.3',
-        trezor2: '2.0.11',
-      },
-      taproot: false,
-      timestamp: false,
-      xprvMagic: 50221816,
-      xpubMagic: 50221772,
-      xpubMagicMultisigSegwitNative: null,
-      xpubMagicMultisigSegwitP2sh: null,
-      xpubMagicSegwitNative: null,
-      xpubMagicSegwitP2sh: null,
-      wif: '',
-    },
-    {
-      addressType: 0,
-      addressTypeP2sh: 5,
-      bech32Prefix: null,
-      blockchainLink: {
-        type: 'blockbook',
-        url: [
-          'https://bch1.trezor.io',
-          'https://bch2.trezor.io',
-          'https://bch3.trezor.io',
-          'https://bch4.trezor.io',
-          'https://bch5.trezor.io',
-        ],
-      },
-      blocktimeSeconds: 600,
-      cashaddrPrefix: 'bitcoincash',
-      coinLabel: 'Bitcoin Cash',
-      coinName: 'Bcash',
-      coinShortcut: 'BCH',
-      curveName: 'secp256k1',
-      decimals: 8,
-      decred: false,
-      defaultFeeB: {
-        economy: 70,
-        high: 200,
-        low: 10,
-        normal: 140,
-      },
-      dustLimit: 546,
-      extraData: false,
-      forceBip143: true,
-      forkId: 0,
-      hashGenesisBlock:
-        '000000000019d6689c085ae165831e934ff763ae46a2a6c172b3f1b60a8ce26f',
-      maxAddressLength: 34,
-      maxfeeKb: 14000000,
-      minAddressLength: 27,
-      minfeeKb: 1000,
-      name: 'Bitcoin Cash',
-      segwit: false,
-      shortcut: 'BCH',
-      signedMessageHeader: 'Bitcoin Signed Message:\n',
-      slip44: 145,
-      support: {
-        connect: true,
-        suite: true,
-        trezor1: '1.6.2',
-        trezor2: '2.0.7',
-      },
-      taproot: false,
-      timestamp: false,
-      xprvMagic: 76066276,
-      xpubMagic: 76067358,
-      xpubMagicMultisigSegwitNative: null,
-      xpubMagicMultisigSegwitP2sh: null,
-      xpubMagicSegwitNative: null,
-      xpubMagicSegwitP2sh: null,
-      wif: '',
-    },
-    {
-      addressType: 111,
-      addressTypeP2sh: 196,
-      bech32Prefix: null,
-      blockchainLink: null,
-      blocktimeSeconds: 600,
-      cashaddrPrefix: 'bchtest',
-      coinLabel: 'Bitcoin Cash Testnet',
-      coinName: 'Bcash Testnet',
-      coinShortcut: 'TBCH',
-      curveName: 'secp256k1',
-      decimals: 8,
-      decred: false,
-      defaultFeeB: {
-        normal: 10,
-      },
-      dustLimit: 546,
-      extraData: false,
-      forceBip143: true,
-      forkId: 0,
-      hashGenesisBlock:
-        '000000000933ea01ad0ee984209779baaec3ced90fa3f408719526f8d77f4943',
-      maxAddressLength: 34,
-      maxfeeKb: 10000000,
-      minAddressLength: 27,
-      minfeeKb: 1000,
-      name: 'Bitcoin Cash Testnet',
-      segwit: false,
-      shortcut: 'TBCH',
-      signedMessageHeader: 'Bitcoin Signed Message:\n',
-      slip44: 1,
-      support: {
-        connect: true,
-        suite: false,
-        trezor1: '1.6.2',
-        trezor2: '2.0.7',
-      },
-      taproot: false,
-      timestamp: false,
-      xprvMagic: 70615956,
-      xpubMagic: 70617039,
-      xpubMagicMultisigSegwitNative: null,
-      xpubMagicMultisigSegwitP2sh: null,
-      xpubMagicSegwitNative: null,
-      xpubMagicSegwitP2sh: null,
-      wif: '',
-    },
-    {
-      addressType: 38,
-      addressTypeP2sh: 23,
-      bech32Prefix: 'btg',
-      blockchainLink: {
-        type: 'blockbook',
-        url: [
-          'https://btg1.trezor.io',
-          'https://btg2.trezor.io',
-          'https://btg3.trezor.io',
-          'https://btg4.trezor.io',
-          'https://btg5.trezor.io',
-        ],
-      },
-      blocktimeSeconds: 600,
-      cashaddrPrefix: null,
-      coinLabel: 'Bitcoin Gold',
-      coinName: 'Bgold',
-      coinShortcut: 'BTG',
-      curveName: 'secp256k1',
-      decimals: 8,
-      decred: false,
-      defaultFeeB: {
-        economy: 70,
-        high: 200,
-        low: 10,
-        normal: 140,
-      },
-      dustLimit: 546,
-      extraData: false,
-      forceBip143: true,
-      forkId: 79,
-      hashGenesisBlock:
-        '000000000019d6689c085ae165831e934ff763ae46a2a6c172b3f1b60a8ce26f',
-      maxAddressLength: 34,
-      maxfeeKb: 380000000,
-      minAddressLength: 27,
-      minfeeKb: 1000,
-      name: 'Bitcoin Gold',
-      segwit: true,
-      shortcut: 'BTG',
-      signedMessageHeader: 'Bitcoin Gold Signed Message:\n',
-      slip44: 156,
-      support: {
-        connect: true,
-        suite: true,
-        trezor1: '1.6.2',
-        trezor2: '2.0.7',
-      },
-      taproot: false,
-      timestamp: false,
-      xprvMagic: 76066276,
-      xpubMagic: 76067358,
-      xpubMagicMultisigSegwitNative: 76067358,
-      xpubMagicMultisigSegwitP2sh: 76067358,
-      xpubMagicSegwitNative: 78792518,
-      xpubMagicSegwitP2sh: 77429938,
-      wif: '',
-    },
-    {
-      addressType: 111,
-      addressTypeP2sh: 196,
-      bech32Prefix: 'tbtg',
-      blockchainLink: null,
-      blocktimeSeconds: 600,
-      cashaddrPrefix: null,
-      coinLabel: 'Bitcoin Gold Testnet',
-      coinName: 'Bgold Testnet',
-      coinShortcut: 'TBTG',
-      curveName: 'secp256k1',
-      decimals: 8,
-      decred: false,
-      defaultFeeB: {
-        economy: 70,
-        high: 200,
-        low: 10,
-        normal: 140,
-      },
-      dustLimit: 546,
-      extraData: false,
-      forceBip143: true,
-      forkId: 79,
-      hashGenesisBlock:
-        '00000000e0781ebe24b91eedc293adfea2f557b53ec379e78959de3853e6f9f6',
-      maxAddressLength: 34,
-      maxfeeKb: 500000,
-      minAddressLength: 27,
-      minfeeKb: 1000,
-      name: 'Bitcoin Gold Testnet',
-      segwit: true,
-      shortcut: 'TBTG',
-      signedMessageHeader: 'Bitcoin Gold Signed Message:\n',
-      slip44: 1,
-      support: {
-        connect: true,
-        suite: false,
-        trezor1: '1.7.1',
-        trezor2: '2.0.8',
-      },
-      taproot: false,
-      timestamp: false,
-      xprvMagic: 70615956,
-      xpubMagic: 70617039,
-      xpubMagicMultisigSegwitNative: 70617039,
-      xpubMagicMultisigSegwitP2sh: 70617039,
-      xpubMagicSegwitNative: 73342198,
-      xpubMagicSegwitP2sh: 71979618,
-      wif: '',
-    },
-    {
-      addressType: 4901,
-      addressTypeP2sh: 5039,
-      bech32Prefix: null,
-      blockchainLink: null,
-      blocktimeSeconds: 150,
-      cashaddrPrefix: null,
-      coinLabel: 'Bitcoin Private',
-      coinName: 'Bprivate',
-      coinShortcut: 'BTCP',
-      curveName: 'secp256k1',
-      decimals: 8,
-      decred: false,
-      defaultFeeB: {
-        normal: 10,
-      },
-      dustLimit: 546,
-      extraData: false,
-      forceBip143: false,
-      forkId: 42,
-      hashGenesisBlock:
-        '0007104ccda289427919efc39dc9e4d499804b7bebc22df55f8b834301260602',
-      maxAddressLength: 95,
-      maxfeeKb: 32000000000,
-      minAddressLength: 35,
-      minfeeKb: 1000,
-      name: 'Bitcoin Private',
-      segwit: false,
-      shortcut: 'BTCP',
-      signedMessageHeader: 'BitcoinPrivate Signed Message:\n',
-      slip44: 183,
-      support: {
-        connect: true,
-        suite: false,
-        trezor1: '1.6.2',
-        trezor2: '2.0.7',
-      },
-      taproot: false,
-      timestamp: false,
-      xprvMagic: 76066276,
-      xpubMagic: 76067358,
-      xpubMagicMultisigSegwitNative: null,
-      xpubMagicMultisigSegwitP2sh: null,
-      xpubMagicSegwitNative: null,
-      xpubMagicSegwitP2sh: null,
-      wif: '',
-    },
-    {
-      addressType: 3,
-      addressTypeP2sh: 125,
-      bech32Prefix: 'btx',
-      blockchainLink: null,
-      blocktimeSeconds: 150,
-      cashaddrPrefix: null,
-      coinLabel: 'Bitcore',
-      coinName: 'Bitcore',
-      coinShortcut: 'BTX',
-      curveName: 'secp256k1',
-      decimals: 8,
-      decred: false,
-      defaultFeeB: {
-        low: 10,
-      },
-      dustLimit: 546,
-      extraData: false,
-      forceBip143: false,
-      forkId: null,
-      hashGenesisBlock:
-        '604148281e5c4b7f2487e5d03cd60d8e6f69411d613f6448034508cea52e9574',
-      maxAddressLength: 34,
-      maxfeeKb: 14000000000,
-      minAddressLength: 27,
-      minfeeKb: 1000,
-      name: 'Bitcore',
-      segwit: true,
-      shortcut: 'BTX',
-      signedMessageHeader: 'BitCore Signed Message:\n',
-      slip44: 160,
-      support: {
-        connect: true,
-        suite: false,
-        trezor1: '1.7.1',
-        trezor2: '2.0.8',
-      },
-      taproot: false,
-      timestamp: false,
-      xprvMagic: 76066276,
-      xpubMagic: 76067358,
-      xpubMagicMultisigSegwitNative: 76067358,
-      xpubMagicMultisigSegwitP2sh: 76067358,
-      xpubMagicSegwitNative: 78792518,
-      xpubMagicSegwitP2sh: 77429938,
-      wif: '',
-    },
-    {
-      addressType: 76,
-      addressTypeP2sh: 16,
-      bech32Prefix: null,
-      blockchainLink: {
-        type: 'blockbook',
-        url: [
-          'https://dash1.trezor.io',
-          'https://dash2.trezor.io',
-          'https://dash3.trezor.io',
-          'https://dash4.trezor.io',
-          'https://dash5.trezor.io',
-        ],
-      },
-      blocktimeSeconds: 150,
-      cashaddrPrefix: null,
-      coinLabel: 'Dash',
-      coinName: 'Dash',
-      coinShortcut: 'DASH',
-      curveName: 'secp256k1',
-      decimals: 8,
-      decred: false,
-      defaultFeeB: {
-        normal: 10,
-      },
-      dustLimit: 5460,
-      extraData: true,
-      forceBip143: false,
-      forkId: null,
-      hashGenesisBlock:
-        '00000ffd590b1485b3caadc19b22e6379c733355108f107a430458cdf3407ab6',
-      maxAddressLength: 34,
-      maxfeeKb: 45000000,
-      minAddressLength: 27,
-      minfeeKb: 1000,
-      name: 'Dash',
-      segwit: false,
-      shortcut: 'DASH',
-      signedMessageHeader: 'DarkCoin Signed Message:\n',
-      slip44: 5,
-      support: {
-        connect: true,
-        suite: true,
-        trezor1: '1.5.2',
-        trezor2: '2.0.5',
-      },
-      taproot: false,
-      timestamp: false,
-      xprvMagic: 50221816,
-      xpubMagic: 50221772,
-      xpubMagicMultisigSegwitNative: null,
-      xpubMagicMultisigSegwitP2sh: null,
-      xpubMagicSegwitNative: null,
-      xpubMagicSegwitP2sh: null,
-      wif: '',
-    },
-    {
-      addressType: 140,
-      addressTypeP2sh: 19,
-      bech32Prefix: null,
-      blockchainLink: null,
-      blocktimeSeconds: 150,
-      cashaddrPrefix: null,
-      coinLabel: 'Dash Testnet',
-      coinName: 'Dash Testnet',
-      coinShortcut: 'tDASH',
-      curveName: 'secp256k1',
-      decimals: 8,
-      decred: false,
-      defaultFeeB: {
-        normal: 10,
-      },
-      dustLimit: 5460,
-      extraData: true,
-      forceBip143: false,
-      forkId: null,
-      hashGenesisBlock:
-        '00000bafbc94add76cb75e2ec92894837288a481e5c005f6563d91623bf8bc2c',
-      maxAddressLength: 34,
-      maxfeeKb: 100000,
-      minAddressLength: 27,
-      minfeeKb: 10000,
-      name: 'Dash Testnet',
-      segwit: false,
-      shortcut: 'tDASH',
-      signedMessageHeader: 'DarkCoin Signed Message:\n',
-      slip44: 1,
-      support: {
-        connect: true,
-        suite: false,
-        trezor1: '1.6.2',
-        trezor2: '2.0.8',
-      },
-      taproot: false,
-      timestamp: false,
-      xprvMagic: 70615956,
-      xpubMagic: 70617039,
-      xpubMagicMultisigSegwitNative: null,
-      xpubMagicMultisigSegwitP2sh: null,
-      xpubMagicSegwitNative: null,
-      xpubMagicSegwitP2sh: null,
-      wif: '',
-    },
-    {
-      addressType: 1855,
-      addressTypeP2sh: 1818,
-      bech32Prefix: null,
-      blockchainLink: null,
-      blocktimeSeconds: 300,
-      cashaddrPrefix: null,
-      coinLabel: 'Decred',
-      coinName: 'Decred',
-      coinShortcut: 'DCR',
-      curveName: 'secp256k1_decred',
-      decimals: 8,
-      decred: true,
-      defaultFeeB: {
-        normal: 10,
-      },
-      dustLimit: 546,
-      extraData: false,
-      forceBip143: false,
-      forkId: null,
-      hashGenesisBlock:
-        '298e5cc3d985bfe7f81dc135f360abe089edd4396b86d2de66b0cef42b21d980',
-      maxAddressLength: 35,
-      maxfeeKb: 220000000,
-      minAddressLength: 35,
-      minfeeKb: 10000,
-      name: 'Decred',
-      segwit: false,
-      shortcut: 'DCR',
-      signedMessageHeader: 'Decred Signed Message:\n',
-      slip44: 42,
-      support: {
-        connect: true,
-        suite: false,
-        trezor1: '1.6.2',
-        trezor2: '2.0.8',
-      },
-      taproot: false,
-      timestamp: false,
-      xprvMagic: 50177256,
-      xpubMagic: 50178342,
-      xpubMagicMultisigSegwitNative: null,
-      xpubMagicMultisigSegwitP2sh: null,
-      xpubMagicSegwitNative: null,
-      xpubMagicSegwitP2sh: null,
-      wif: '',
-    },
-    {
-      addressType: 3873,
-      addressTypeP2sh: 3836,
-      bech32Prefix: null,
-      blockchainLink: null,
-      blocktimeSeconds: 120,
-      cashaddrPrefix: null,
-      coinLabel: 'Decred Testnet',
-      coinName: 'Decred Testnet',
-      coinShortcut: 'TDCR',
-      curveName: 'secp256k1_decred',
-      decimals: 8,
-      decred: true,
-      defaultFeeB: {
-        normal: 10,
-      },
-      dustLimit: 546,
-      extraData: false,
-      forceBip143: false,
-      forkId: null,
-      hashGenesisBlock:
-        'a649dce53918caf422e9c711c858837e08d626ecfcd198969b24f7b634a49bac',
-      maxAddressLength: 35,
-      maxfeeKb: 10000000,
-      minAddressLength: 35,
-      minfeeKb: 1000,
-      name: 'Decred Testnet',
-      segwit: false,
-      shortcut: 'TDCR',
-      signedMessageHeader: 'Decred Signed Message:\n',
-      slip44: 1,
-      support: {
-        connect: true,
-        suite: false,
-        trezor1: '1.6.2',
-        trezor2: '2.0.8',
-      },
-      taproot: false,
-      timestamp: false,
-      xprvMagic: 70615959,
-      xpubMagic: 70617041,
-      xpubMagicMultisigSegwitNative: null,
-      xpubMagicMultisigSegwitP2sh: null,
-      xpubMagicSegwitNative: null,
-      xpubMagicSegwitP2sh: null,
-      wif: '',
-    },
-    {
-      addressType: 30,
-      addressTypeP2sh: 63,
-      bech32Prefix: 'dgb',
-      blockchainLink: {
-        type: 'blockbook',
-        url: ['https://dgb1.trezor.io', 'https://dgb2.trezor.io'],
-      },
-      blocktimeSeconds: 15,
-      cashaddrPrefix: null,
-      coinLabel: 'DigiByte',
-      coinName: 'DigiByte',
-      coinShortcut: 'DGB',
-      curveName: 'secp256k1',
-      decimals: 8,
-      decred: false,
-      defaultFeeB: {
-        economy: 70,
-        high: 200,
-        low: 10,
-        normal: 140,
-      },
-      dustLimit: 546,
-      extraData: false,
-      forceBip143: false,
-      forkId: null,
-      hashGenesisBlock:
-        '7497ea1b465eb39f1c8f507bc877078fe016d6fcb6dfad3a64c98dcc6e1e8496',
-      maxAddressLength: 34,
-      maxfeeKb: 130000000000,
-      minAddressLength: 27,
-      minfeeKb: 1000,
-      name: 'DigiByte',
-      segwit: true,
-      shortcut: 'DGB',
-      signedMessageHeader: 'DigiByte Signed Message:\n',
-      slip44: 20,
-      support: {
-        connect: true,
-        suite: true,
-        trezor1: '1.6.3',
-        trezor2: '2.0.7',
-      },
-      taproot: false,
-      timestamp: false,
-      xprvMagic: 76066276,
-      xpubMagic: 76067358,
-      xpubMagicMultisigSegwitNative: 76067358,
-      xpubMagicMultisigSegwitP2sh: 76067358,
-      xpubMagicSegwitNative: 78792518,
-      xpubMagicSegwitP2sh: 77429938,
-      wif: '',
-    },
-    {
-      addressType: 30,
-      addressTypeP2sh: 22,
-      bech32Prefix: null,
-      blockchainLink: {
-        type: 'blockbook',
-        url: [
-          'https://doge1.trezor.io',
-          'https://doge2.trezor.io',
-          'https://doge3.trezor.io',
-          'https://doge4.trezor.io',
-          'https://doge5.trezor.io',
-        ],
-      },
-      blocktimeSeconds: 60,
-      cashaddrPrefix: null,
-      coinLabel: 'Dogecoin',
-      coinName: 'Dogecoin',
-      coinShortcut: 'DOGE',
-      curveName: 'secp256k1',
-      decimals: 8,
-      decred: false,
-      defaultFeeB: {
-        normal: 1000,
-      },
-      dustLimit: 999999,
-      extraData: false,
-      forceBip143: false,
-      forkId: null,
-      hashGenesisBlock:
-        '1a91e3dace36e2be3bf030a65679fe821aa1d6ef92e7c9902eb318182c355691',
-      maxAddressLength: 34,
-      maxfeeKb: 1200000000000,
-      minAddressLength: 27,
-      minfeeKb: 100000,
-      name: 'Dogecoin Testnet',
-      segwit: false,
-      shortcut: 'DOGE',
-      signedMessageHeader: 'Dogecoin Signed Message:\n',
-      slip44: 3,
-      support: {
-        connect: true,
-        suite: true,
-        trezor1: '1.5.2',
-        trezor2: '2.0.5',
-      },
-      taproot: false,
-      timestamp: false,
-      xprvMagic: 49988504,
-      xpubMagic: 49990397,
-      xpubMagicMultisigSegwitNative: null,
-      xpubMagicMultisigSegwitP2sh: null,
-      xpubMagicSegwitNative: null,
-      xpubMagicSegwitP2sh: null,
-      wif: '',
-    },
-    {
-      addressType: 14,
-      addressTypeP2sh: 5,
-      bech32Prefix: 'fc',
-      blockchainLink: null,
-      blocktimeSeconds: 60,
-      cashaddrPrefix: null,
-      coinLabel: 'Feathercoin',
-      coinName: 'Feathercoin',
-      coinShortcut: 'FTC',
-      curveName: 'secp256k1',
-      decimals: 8,
-      decred: false,
-      defaultFeeB: {
-        normal: 1000,
-      },
-      dustLimit: 54600,
-      extraData: false,
-      forceBip143: false,
-      forkId: null,
-      hashGenesisBlock:
-        '12a765e31ffd4059bada1e25190f6e98c99d9714d334efa41a195a7e7e04bfe2',
-      maxAddressLength: 34,
-      maxfeeKb: 390000000000,
-      minAddressLength: 27,
-      minfeeKb: 1000,
-      name: 'Feathercoin',
-      segwit: true,
-      shortcut: 'FTC',
-      signedMessageHeader: 'Feathercoin Signed Message:\n',
-      slip44: 8,
-      support: {
-        connect: true,
-        suite: false,
-        trezor1: '1.7.1',
-        trezor2: '2.0.8',
-      },
-      taproot: false,
-      timestamp: false,
-      xprvMagic: 76077806,
-      xpubMagic: 76069926,
-      xpubMagicMultisigSegwitNative: 76069926,
-      xpubMagicMultisigSegwitP2sh: 76069926,
-      xpubMagicSegwitNative: 78792518,
-      xpubMagicSegwitP2sh: 77429938,
-      wif: '',
-    },
-    {
-      addressType: 82,
-      addressTypeP2sh: 7,
-      bech32Prefix: null,
-      blockchainLink: null,
-      blocktimeSeconds: 600,
-      cashaddrPrefix: null,
-      coinLabel: 'Firo',
-      coinName: 'Firo',
-      coinShortcut: 'FIRO',
-      curveName: 'secp256k1',
-      decimals: 8,
-      decred: false,
-      defaultFeeB: {
-        economy: 10,
-        high: 200,
-        low: 1,
-        normal: 100,
-      },
-      dustLimit: 546,
-      extraData: true,
-      forceBip143: false,
-      forkId: null,
-      hashGenesisBlock:
-        '4381deb85b1b2c9843c222944b616d997516dcbd6a964e1eaf0def0830695233',
-      maxAddressLength: 34,
-      maxfeeKb: 640000000,
-      minAddressLength: 27,
-      minfeeKb: 0,
-      name: 'Firo',
-      segwit: false,
-      shortcut: 'FIRO',
-      signedMessageHeader: 'Zcoin Signed Message:\n',
-      slip44: 136,
-      support: {
-        connect: true,
-        suite: false,
-        trezor1: '1.6.2',
-        trezor2: '2.0.7',
-      },
-      taproot: false,
-      timestamp: false,
-      xprvMagic: 76066276,
-      xpubMagic: 76067358,
-      xpubMagicMultisigSegwitNative: null,
-      xpubMagicMultisigSegwitP2sh: null,
-      xpubMagicSegwitNative: null,
-      xpubMagicSegwitP2sh: null,
-      wif: '',
-    },
-    {
-      addressType: 65,
-      addressTypeP2sh: 178,
-      bech32Prefix: null,
-      blockchainLink: null,
-      blocktimeSeconds: 600,
-      cashaddrPrefix: null,
-      coinLabel: 'Firo Testnet',
-      coinName: 'Firo Testnet',
-      coinShortcut: 'tFIRO',
-      curveName: 'secp256k1',
-      decimals: 8,
-      decred: false,
-      defaultFeeB: {
-        economy: 10,
-        high: 200,
-        low: 1,
-        normal: 100,
-      },
-      dustLimit: 546,
-      extraData: true,
-      forceBip143: false,
-      forkId: null,
-      hashGenesisBlock:
-        '7ac038c193c2158c428c59f9ae0c02a07115141c6e9dc244ae96132e99b4e642',
-      maxAddressLength: 35,
-      maxfeeKb: 1000000,
-      minAddressLength: 27,
-      minfeeKb: 0,
-      name: 'Firo Testnet',
-      segwit: false,
-      shortcut: 'tFIRO',
-      signedMessageHeader: 'Zcoin Signed Message:\n',
-      slip44: 1,
-      support: {
-        connect: true,
-        suite: false,
-        trezor1: '1.6.2',
-        trezor2: '2.0.7',
-      },
-      taproot: false,
-      timestamp: false,
-      xprvMagic: 70615956,
-      xpubMagic: 70617039,
-      xpubMagicMultisigSegwitNative: null,
-      xpubMagicMultisigSegwitP2sh: null,
-      xpubMagicSegwitNative: null,
-      xpubMagicSegwitP2sh: null,
-      wif: '',
-    },
-    {
-      addressType: 35,
-      addressTypeP2sh: 94,
-      bech32Prefix: 'flo',
-      blockchainLink: null,
-      blocktimeSeconds: 40,
-      cashaddrPrefix: null,
-      coinLabel: 'Flo',
-      coinName: 'Florincoin',
-      coinShortcut: 'FLO',
-      curveName: 'secp256k1',
-      decimals: 8,
-      decred: false,
-      defaultFeeB: {
-        normal: 1000,
-      },
-      dustLimit: 546,
-      extraData: true,
-      forceBip143: false,
-      forkId: null,
-      hashGenesisBlock:
-        '09c7781c9df90708e278c35d38ea5c9041d7ecfcdd1c56ba67274b7cff3e1cea',
-      maxAddressLength: 34,
-      maxfeeKb: 78000000000,
-      minAddressLength: 27,
-      minfeeKb: 100000,
-      name: 'Flo',
-      segwit: true,
-      shortcut: 'FLO',
-      signedMessageHeader: 'Florincoin Signed Message:\n',
-      slip44: 216,
-      support: {
-        connect: true,
-        suite: false,
-        trezor1: '1.7.2',
-        trezor2: '2.0.11',
-      },
-      taproot: false,
-      timestamp: false,
-      xprvMagic: 15264107,
-      xpubMagic: 1526049,
-      xpubMagicMultisigSegwitNative: 1526049,
-      xpubMagicMultisigSegwitP2sh: 1526049,
-      xpubMagicSegwitNative: 78792518,
-      xpubMagicSegwitP2sh: 28471030,
-      wif: '',
-    },
-    {
-      addressType: 36,
-      addressTypeP2sh: 16,
-      bech32Prefix: 'fc',
-      blockchainLink: {
-        type: 'blockbook',
-        url: ['https://explorer.fujicoin.org'],
-      },
-      blocktimeSeconds: 60,
-      cashaddrPrefix: null,
-      coinLabel: 'Fujicoin',
-      coinName: 'Fujicoin',
-      coinShortcut: 'FJC',
-      curveName: 'secp256k1',
-      decimals: 8,
-      decred: false,
-      defaultFeeB: {
-        economy: 20000,
-        high: 100000,
-        low: 10000,
-        normal: 50000,
-      },
-      dustLimit: 546,
-      extraData: false,
-      forceBip143: false,
-      forkId: null,
-      hashGenesisBlock:
-        'adb6d9cfd74075e7f91608add4bd2a2ea636f70856183086842667a1597714a0',
-      maxAddressLength: 34,
-      maxfeeKb: 35000000000000,
-      minAddressLength: 27,
-      minfeeKb: 10000000,
-      name: 'Fujicoin',
-      segwit: true,
-      shortcut: 'FJC',
-      signedMessageHeader: 'FujiCoin Signed Message:\n',
-      slip44: 75,
-      support: {
-        connect: true,
-        suite: false,
-        trezor1: '1.6.1',
-        trezor2: '2.0.5',
-      },
-      taproot: true,
-      timestamp: false,
-      xprvMagic: 76066276,
-      xpubMagic: 76067358,
-      xpubMagicMultisigSegwitNative: 44728019,
-      xpubMagicMultisigSegwitP2sh: 43365439,
-      xpubMagicSegwitNative: 78792518,
-      xpubMagicSegwitP2sh: 77429938,
-      wif: '',
-    },
-    {
-      addressType: 38,
-      addressTypeP2sh: 62,
-      bech32Prefix: 'game',
-      blockchainLink: {
-        type: 'blockbook',
-        url: ['https://blockbook.gamecredits.network'],
-      },
-      blocktimeSeconds: 90,
-      cashaddrPrefix: null,
-      coinLabel: 'GameCredits',
-      coinName: 'GameCredits',
-      coinShortcut: 'GAME',
-      curveName: 'secp256k1',
-      decimals: 8,
-      decred: false,
-      defaultFeeB: {
-        normal: 1000,
-      },
-      dustLimit: 54600,
-      extraData: false,
-      forceBip143: false,
-      forkId: null,
-      hashGenesisBlock:
-        '91ec5f25ee9a0ffa1af7d4da4db9a552228dd2dc77cdb15b738be4e1f55f30ee',
-      maxAddressLength: 34,
-      maxfeeKb: 53000000000,
-      minAddressLength: 27,
-      minfeeKb: 100000,
-      name: 'GameCredits',
-      segwit: true,
-      shortcut: 'GAME',
-      signedMessageHeader: 'GameCredits Signed Message:\n',
-      slip44: 101,
-      support: {
-        connect: true,
-        suite: false,
-        trezor1: '1.7.1',
-        trezor2: '2.0.8',
-      },
-      taproot: false,
-      timestamp: false,
-      xprvMagic: 27108450,
-      xpubMagic: 27106558,
-      xpubMagicMultisigSegwitNative: 27108450,
-      xpubMagicMultisigSegwitP2sh: 27108450,
-      xpubMagicSegwitNative: 78792518,
-      xpubMagicSegwitP2sh: 28471030,
-      wif: '',
-    },
-    {
-      addressType: 60,
-      addressTypeP2sh: 85,
-      bech32Prefix: null,
-      blockchainLink: null,
-      blocktimeSeconds: 60,
-      cashaddrPrefix: null,
-      coinLabel: 'Komodo',
-      coinName: 'Komodo',
-      coinShortcut: 'KMD',
-      curveName: 'secp256k1',
-      decimals: 8,
-      decred: false,
-      defaultFeeB: {
-        normal: 10,
-      },
-      dustLimit: 546,
-      extraData: true,
-      forceBip143: false,
-      forkId: null,
-      hashGenesisBlock:
-        '027e3758c3a65b12aa1046462b486d0a63bfa1beae327897f56c5cfb7daaae71',
-      maxAddressLength: 34,
-      maxfeeKb: 4800000000,
-      minAddressLength: 27,
-      minfeeKb: 1000,
-      name: 'Komodo',
-      segwit: false,
-      shortcut: 'KMD',
-      signedMessageHeader: 'Komodo Signed Message:\n',
-      slip44: 141,
-      support: {
-        connect: true,
-        suite: false,
-        trezor1: '1.8.0',
-        trezor2: '2.0.11',
-      },
-      taproot: false,
-      timestamp: false,
-      xprvMagic: 76066276,
-      xpubMagic: 76067358,
-      xpubMagicMultisigSegwitNative: null,
-      xpubMagicMultisigSegwitP2sh: null,
-      xpubMagicSegwitNative: null,
-      xpubMagicSegwitP2sh: null,
-      wif: '',
-    },
-    {
-      addressType: 6198,
-      addressTypeP2sh: 6203,
-      bech32Prefix: null,
-      blockchainLink: {
-        type: 'blockbook',
-        url: ['https://blockbook.kotocoin.info'],
-      },
-      blocktimeSeconds: 60,
-      cashaddrPrefix: null,
-      coinLabel: 'Koto',
-      coinName: 'Koto',
-      coinShortcut: 'KOTO',
-      curveName: 'secp256k1',
-      decimals: 8,
-      decred: false,
-      defaultFeeB: {
-        normal: 10,
-      },
-      dustLimit: 546,
-      extraData: true,
-      forceBip143: false,
-      forkId: null,
-      hashGenesisBlock:
-        '6d424c350729ae633275d51dc3496e16cd1b1d195c164da00f39c499a2e9959e',
-      maxAddressLength: 95,
-      maxfeeKb: 1000000,
-      minAddressLength: 35,
-      minfeeKb: 1000,
-      name: 'Koto',
-      segwit: false,
-      shortcut: 'KOTO',
-      signedMessageHeader: 'Koto Signed Message:\n',
-      slip44: 510,
-      support: {
-        connect: true,
-        suite: false,
-        trezor1: '1.7.1',
-        trezor2: '2.0.8',
-      },
-      taproot: false,
-      timestamp: false,
-      xprvMagic: 76066276,
-      xpubMagic: 76067358,
-      xpubMagicMultisigSegwitNative: null,
-      xpubMagicMultisigSegwitP2sh: null,
-      xpubMagicSegwitNative: null,
-      xpubMagicSegwitP2sh: null,
-      wif: '',
-    },
-    {
-      addressType: 48,
-      addressTypeP2sh: 5,
-      bech32Prefix: 'ltc',
-      blockchainLink: {
-        type: 'blockbook',
-        url: [
-          'https://ltc1.trezor.io',
-          'https://ltc2.trezor.io',
-          'https://ltc3.trezor.io',
-          'https://ltc4.trezor.io',
-          'https://ltc5.trezor.io',
-        ],
-      },
-      blocktimeSeconds: 150,
-      cashaddrPrefix: null,
-      coinLabel: 'Litecoin',
-      coinName: 'Litecoin',
-      coinShortcut: 'LTC',
-      curveName: 'secp256k1',
-      decimals: 8,
-      decred: false,
-      defaultFeeB: {
-        normal: 1000,
-      },
-      dustLimit: 546,
-      extraData: false,
-      forceBip143: false,
-      forkId: null,
-      hashGenesisBlock:
-        '12a765e31ffd4059bada1e25190f6e98c99d9714d334efa41a195a7e7e04bfe2',
-      maxAddressLength: 34,
-      maxfeeKb: 67000000,
-      minAddressLength: 27,
-      minfeeKb: 1000,
-      name: 'Litecoin',
-      segwit: true,
-      shortcut: 'LTC',
-      signedMessageHeader: 'Litecoin Signed Message:\n',
-      slip44: 2,
-      support: {
-        connect: true,
-        suite: true,
-        trezor1: '1.5.2',
-        trezor2: '2.0.5',
-      },
-      taproot: false,
-      timestamp: false,
-      xprvMagic: 27106558,
-      xpubMagic: 27108450,
-      xpubMagicMultisigSegwitNative: 27108450,
-      xpubMagicMultisigSegwitP2sh: 27108450,
-      xpubMagicSegwitNative: 78792518,
-      xpubMagicSegwitP2sh: 28471030,
-      wif: '',
-    },
-    {
-      addressType: 111,
-      addressTypeP2sh: 196,
-      bech32Prefix: 'tltc',
-      blockchainLink: null,
-      blocktimeSeconds: 150,
-      cashaddrPrefix: null,
-      coinLabel: 'Litecoin Testnet',
-      coinName: 'Litecoin Testnet',
-      coinShortcut: 'tLTC',
-      curveName: 'secp256k1',
-      decimals: 8,
-      decred: false,
-      defaultFeeB: {
-        normal: 10,
-      },
-      dustLimit: 54600,
-      extraData: false,
-      forceBip143: false,
-      forkId: null,
-      hashGenesisBlock:
-        '4966625a4b2851d9fdee139e56211a0d88575f59ed816ff5e6a63deb4e3e29a0',
-      maxAddressLength: 34,
-      maxfeeKb: 40000000,
-      minAddressLength: 27,
-      minfeeKb: 1000,
-      name: 'Litecoin Testnet',
-      segwit: true,
-      shortcut: 'tLTC',
-      signedMessageHeader: 'Litecoin Signed Message:\n',
-      slip44: 1,
-      support: {
-        connect: true,
-        suite: false,
-        trezor1: '1.6.2',
-        trezor2: '2.0.7',
-      },
-      taproot: false,
-      timestamp: false,
-      xprvMagic: 70615956,
-      xpubMagic: 70617039,
-      xpubMagicMultisigSegwitNative: 70617039,
-      xpubMagicMultisigSegwitP2sh: 70617039,
-      xpubMagicSegwitNative: 73342198,
-      xpubMagicSegwitP2sh: 71979618,
-      wif: '',
-    },
-    {
-      addressType: 50,
-      addressTypeP2sh: 55,
-      bech32Prefix: 'mona',
-      blockchainLink: {
-        type: 'blockbook',
-        url: ['https://blockbook.electrum-mona.org'],
-      },
-      blocktimeSeconds: 90,
-      cashaddrPrefix: null,
-      coinLabel: 'Monacoin',
-      coinName: 'Monacoin',
-      coinShortcut: 'MONA',
-      curveName: 'secp256k1',
-      decimals: 8,
-      decred: false,
-      defaultFeeB: {
-        normal: 1000,
-      },
-      dustLimit: 54600,
-      extraData: false,
-      forceBip143: false,
-      forkId: null,
-      hashGenesisBlock:
-        'ff9f1c0116d19de7c9963845e129f9ed1bfc0b376eb54fd7afa42e0d418c8bb6',
-      maxAddressLength: 34,
-      maxfeeKb: 2100000000,
-      minAddressLength: 27,
-      minfeeKb: 100000,
-      name: 'Monacoin',
-      segwit: true,
-      shortcut: 'MONA',
-      signedMessageHeader: 'Monacoin Signed Message:\n',
-      slip44: 22,
-      support: {
-        connect: true,
-        suite: false,
-        trezor1: '1.6.0',
-        trezor2: '2.0.5',
-      },
-      taproot: false,
-      timestamp: false,
-      xprvMagic: 76066276,
-      xpubMagic: 76067358,
-      xpubMagicMultisigSegwitNative: 76067358,
-      xpubMagicMultisigSegwitP2sh: 76067358,
-      xpubMagicSegwitNative: 78792518,
-      xpubMagicSegwitP2sh: 77429938,
-      wif: '',
-    },
-    {
-      addressType: 16,
-      addressTypeP2sh: 76,
-      bech32Prefix: null,
-      blockchainLink: {
-        type: 'blockbook',
-        url: ['https://blockbook.monetaryunit.org'],
-      },
-      blocktimeSeconds: 40,
-      cashaddrPrefix: null,
-      coinLabel: 'MonetaryUnit',
-      coinName: 'MonetaryUnit',
-      coinShortcut: 'MUE',
-      curveName: 'secp256k1',
-      decimals: 8,
-      decred: false,
-      defaultFeeB: {
-        normal: 10,
-      },
-      dustLimit: 5460,
-      extraData: false,
-      forceBip143: false,
-      forkId: null,
-      hashGenesisBlock:
-        '0b58ed450b3819ca54ab0054c4d220ca4f887d21c9e55d2a333173adf76d987f',
-      maxAddressLength: 34,
-      maxfeeKb: 600000000000,
-      minAddressLength: 27,
-      minfeeKb: 1000,
-      name: 'MonetaryUnit Testnet',
-      segwit: false,
-      shortcut: 'MUE',
-      signedMessageHeader: 'MonetaryUnit Signed Message:\n',
-      slip44: 31,
-      support: {
-        connect: true,
-        suite: false,
-        trezor1: '1.7.1',
-        trezor2: '2.0.8',
-      },
-      taproot: false,
-      timestamp: false,
-      xprvMagic: 76066276,
-      xpubMagic: 76067358,
-      xpubMagicMultisigSegwitNative: null,
-      xpubMagicMultisigSegwitP2sh: null,
-      xpubMagicSegwitNative: null,
-      xpubMagicSegwitP2sh: null,
-      wif: '',
-    },
-    {
-      addressType: 38,
-      addressTypeP2sh: 53,
-      bech32Prefix: 'nix',
-      blockchainLink: null,
-      blocktimeSeconds: 120,
-      cashaddrPrefix: null,
-      coinLabel: 'NIX',
-      coinName: 'NIX',
-      coinShortcut: 'NIX',
-      curveName: 'secp256k1',
-      decimals: 8,
-      decred: false,
-      defaultFeeB: {
-        normal: 1000,
-      },
-      dustLimit: 54600,
-      extraData: false,
-      forceBip143: false,
-      forkId: null,
-      hashGenesisBlock:
-        'dd28ad86def767c3cfc34267a950d871fc7462bc57ea4a929fc3596d9b598e41',
-      maxAddressLength: 34,
-      maxfeeKb: 30000000000,
-      minAddressLength: 27,
-      minfeeKb: 0,
-      name: 'NIX',
-      segwit: true,
-      shortcut: 'NIX',
-      signedMessageHeader: 'NIX Signed Message:\n',
-      slip44: 400,
-      support: {
-        connect: true,
-        suite: false,
-        trezor1: '1.7.2',
-        trezor2: '2.0.11',
-      },
-      taproot: false,
-      timestamp: false,
-      xprvMagic: 76066276,
-      xpubMagic: 76067358,
-      xpubMagicMultisigSegwitNative: 76067358,
-      xpubMagicMultisigSegwitP2sh: 76067358,
-      xpubMagicSegwitNative: 78792518,
-      xpubMagicSegwitP2sh: 77429938,
-      wif: '',
-    },
-    {
-      addressType: 52,
-      addressTypeP2sh: 5,
-      bech32Prefix: null,
-      blockchainLink: {
-        type: 'blockbook',
-        url: ['https://nmc1.trezor.io', 'https://nmc2.trezor.io'],
-      },
-      blocktimeSeconds: 600,
-      cashaddrPrefix: null,
-      coinLabel: 'Namecoin',
-      coinName: 'Namecoin',
-      coinShortcut: 'NMC',
-      curveName: 'secp256k1',
-      decimals: 8,
-      decred: false,
-      defaultFeeB: {
-        normal: 10,
-      },
-      dustLimit: 2940,
-      extraData: false,
-      forceBip143: false,
-      forkId: null,
-      hashGenesisBlock:
-        '000000000062b72c5e2ceb45fbc8587e807c155b0da735e6483dfba2f0a9c770',
-      maxAddressLength: 34,
-      maxfeeKb: 8700000000,
-      minAddressLength: 27,
-      minfeeKb: 100000,
-      name: 'Namecoin',
-      segwit: false,
-      shortcut: 'NMC',
-      signedMessageHeader: 'Namecoin Signed Message:\n',
-      slip44: 7,
-      support: {
-        connect: true,
-        suite: true,
-        trezor1: '1.5.2',
-        trezor2: '2.0.5',
-      },
-      taproot: false,
-      timestamp: false,
-      xprvMagic: 76066276,
-      xpubMagic: 76067358,
-      xpubMagicMultisigSegwitNative: null,
-      xpubMagicMultisigSegwitP2sh: null,
-      xpubMagicSegwitNative: null,
-      xpubMagicSegwitP2sh: null,
-      wif: '',
-    },
-    {
-      addressType: 55,
-      addressTypeP2sh: 117,
-      bech32Prefix: 'pc',
-      blockchainLink: {
-        type: 'blockbook',
-        url: ['https://blockbook.peercoin.net'],
-      },
-      blocktimeSeconds: 600,
-      cashaddrPrefix: null,
-      coinLabel: 'Peercoin',
-      coinName: 'Peercoin',
-      coinShortcut: 'PPC',
-      curveName: 'secp256k1',
-      decimals: 6,
-      decred: false,
-      defaultFeeB: {
-        normal: 10,
-      },
-      dustLimit: 10000,
-      extraData: false,
-      forceBip143: false,
-      forkId: null,
-      hashGenesisBlock:
-        '0000000032fe677166d54963b62a4677d8957e87c508eaa4fd7eb1c880cd27e3',
-      maxAddressLength: 34,
-      maxfeeKb: 13000000000,
-      minAddressLength: 27,
-      minfeeKb: 10000,
-      name: 'Peercoin',
-      segwit: true,
-      shortcut: 'PPC',
-      signedMessageHeader: 'Peercoin Signed Message:\n',
-      slip44: 6,
-      support: {
-        connect: true,
-        suite: false,
-        trezor1: '1.8.4',
-        trezor2: '2.1.9',
-      },
-      taproot: false,
-      timestamp: true,
-      xprvMagic: 76066276,
-      xpubMagic: 76067358,
-      xpubMagicMultisigSegwitNative: 76067358,
-      xpubMagicMultisigSegwitP2sh: 76067358,
-      xpubMagicSegwitNative: 78792518,
-      xpubMagicSegwitP2sh: 77429938,
-      wif: '',
-    },
-    {
-      addressType: 111,
-      addressTypeP2sh: 196,
-      bech32Prefix: 'tpc',
-      blockchainLink: {
-        type: 'blockbook',
-        url: ['https://tblockbook.peercoin.net'],
-      },
-      blocktimeSeconds: 600,
-      cashaddrPrefix: null,
-      coinLabel: 'Peercoin Testnet',
-      coinName: 'Peercoin Testnet',
-      coinShortcut: 'tPPC',
-      curveName: 'secp256k1',
-      decimals: 6,
-      decred: false,
-      defaultFeeB: {
-        normal: 10,
-      },
-      dustLimit: 10000,
-      extraData: false,
-      forceBip143: false,
-      forkId: null,
-      hashGenesisBlock:
-        '00000001f757bb737f6596503e17cd17b0658ce630cc727c0cca81aec47c9f06',
-      maxAddressLength: 34,
-      maxfeeKb: 2000000,
-      minAddressLength: 27,
-      minfeeKb: 10000,
-      name: 'Peercoin Testnet',
-      segwit: true,
-      shortcut: 'tPPC',
-      signedMessageHeader: 'Peercoin Signed Message:\n',
-      slip44: 1,
-      support: {
-        connect: true,
-        suite: false,
-        trezor1: '1.8.4',
-        trezor2: '2.1.9',
-      },
-      taproot: false,
-      timestamp: true,
-      xprvMagic: 70615956,
-      xpubMagic: 70617039,
-      xpubMagicMultisigSegwitNative: 70617039,
-      xpubMagicMultisigSegwitP2sh: 70617039,
-      xpubMagicSegwitNative: 73342198,
-      xpubMagicSegwitP2sh: 71979618,
-      wif: '',
-    },
-    {
-      addressType: 55,
-      addressTypeP2sh: 56,
-      bech32Prefix: null,
-      blockchainLink: {
-        type: 'blockbook',
-        url: ['https://blockbook.polispay.org'],
-      },
-      blocktimeSeconds: 120,
-      cashaddrPrefix: null,
-      coinLabel: 'Polis',
-      coinName: 'Polis',
-      coinShortcut: 'POLIS',
-      curveName: 'secp256k1',
-      decimals: 8,
-      decred: false,
-      defaultFeeB: {
-        normal: 10,
-      },
-      dustLimit: 5460,
-      extraData: false,
-      forceBip143: false,
-      forkId: null,
-      hashGenesisBlock:
-        '000009701eb781a8113b1af1d814e2f060f6408a2c990db291bc5108a1345c1e',
-      maxAddressLength: 34,
-      maxfeeKb: 4900000000,
-      minAddressLength: 27,
-      minfeeKb: 1000,
-      name: 'Polis',
-      segwit: false,
-      shortcut: 'POLIS',
-      signedMessageHeader: 'Polis Signed Message:\n',
-      slip44: 1997,
-      support: {
-        connect: true,
-        suite: false,
-        trezor1: '1.8.2',
-        trezor2: '2.1.1',
-      },
-      taproot: false,
-      timestamp: false,
-      xprvMagic: 65165637,
-      xpubMagic: 65166718,
-      xpubMagicMultisigSegwitNative: null,
-      xpubMagicMultisigSegwitP2sh: null,
-      xpubMagicSegwitNative: null,
-      xpubMagicSegwitP2sh: null,
-      wif: '',
-    },
-    {
-      addressType: 23,
-      addressTypeP2sh: 83,
-      bech32Prefix: null,
-      blockchainLink: null,
-      blocktimeSeconds: 60,
-      cashaddrPrefix: null,
-      coinLabel: 'Primecoin',
-      coinName: 'Primecoin',
-      coinShortcut: 'XPM',
-      curveName: 'secp256k1',
-      decimals: 8,
-      decred: false,
-      defaultFeeB: {
-        normal: 10,
-      },
-      dustLimit: 546,
-      extraData: false,
-      forceBip143: false,
-      forkId: null,
-      hashGenesisBlock:
-        '963d17ba4dc753138078a2f56afb3af9674e2546822badff26837db9a0152106',
-      maxAddressLength: 35,
-      maxfeeKb: 89000000000,
-      minAddressLength: 26,
-      minfeeKb: 1000,
-      name: 'Primecoin',
-      segwit: false,
-      shortcut: 'XPM',
-      signedMessageHeader: 'Primecoin Signed Message:\n',
-      slip44: 24,
-      support: {
-        connect: true,
-        suite: false,
-        trezor1: '1.8.0',
-        trezor2: '2.0.11',
-      },
-      taproot: false,
-      timestamp: false,
-      xprvMagic: 76066276,
-      xpubMagic: 76067358,
-      xpubMagicMultisigSegwitNative: null,
-      xpubMagicMultisigSegwitP2sh: null,
-      xpubMagicSegwitNative: null,
-      xpubMagicSegwitP2sh: null,
-      wif: '',
-    },
-    {
-      addressType: 60,
-      addressTypeP2sh: 122,
-      bech32Prefix: null,
-      blockchainLink: {
-        type: 'blockbook',
-        url: ['https://blockbook.ravencoin.org'],
-      },
-      blocktimeSeconds: 60,
-      cashaddrPrefix: null,
-      coinLabel: 'Ravencoin',
-      coinName: 'Ravencoin',
-      coinShortcut: 'RVN',
-      curveName: 'secp256k1',
-      decimals: 8,
-      decred: false,
-      defaultFeeB: {
-        low: 10,
-      },
-      dustLimit: 546,
-      extraData: false,
-      forceBip143: false,
-      forkId: null,
-      hashGenesisBlock:
-        '0000006b444bc2f2ffe627be9d9e7e7a0730000870ef6eb6da46c8eae389df90',
-      maxAddressLength: 34,
-      maxfeeKb: 170000000000,
-      minAddressLength: 27,
-      minfeeKb: 1000,
-      name: 'Ravencoin',
-      segwit: false,
-      shortcut: 'RVN',
-      signedMessageHeader: 'Raven Signed Message:\n',
-      slip44: 175,
-      support: {
-        connect: true,
-        suite: false,
-        trezor1: '1.7.2',
-        trezor2: '2.0.10',
-      },
-      taproot: false,
-      timestamp: false,
-      xprvMagic: 76066276,
-      xpubMagic: 76067358,
-      xpubMagicMultisigSegwitNative: null,
-      xpubMagicMultisigSegwitP2sh: null,
-      xpubMagicSegwitNative: null,
-      xpubMagicSegwitP2sh: null,
-      wif: '',
-    },
-    {
-      addressType: 25,
-      addressTypeP2sh: 105,
-      bech32Prefix: null,
-      blockchainLink: {
-        type: 'blockbook',
-        url: ['https://blockbook.ritocoin.org'],
-      },
-      blocktimeSeconds: 60,
-      cashaddrPrefix: null,
-      coinLabel: 'Ritocoin',
-      coinName: 'Ritocoin',
-      coinShortcut: 'RITO',
-      curveName: 'secp256k1',
-      decimals: 8,
-      decred: false,
-      defaultFeeB: {
-        low: 10,
-      },
-      dustLimit: 546,
-      extraData: false,
-      forceBip143: false,
-      forkId: null,
-      hashGenesisBlock:
-        '00000075e344bdf1c0e433f453764b1830a7aa19b2a5213e707502a22b779c1b',
-      maxAddressLength: 34,
-      maxfeeKb: 39000000000000,
-      minAddressLength: 27,
-      minfeeKb: 1000,
-      name: 'Ritocoin',
-      segwit: false,
-      shortcut: 'RITO',
-      signedMessageHeader: 'Rito Signed Message:\n',
-      slip44: 19169,
-      support: {
-        connect: true,
-        suite: false,
-        trezor1: '1.8.2',
-        trezor2: '2.1.1',
-      },
-      taproot: false,
-      timestamp: false,
-      xprvMagic: 87326380,
-      xpubMagic: 87353290,
-      xpubMagicMultisigSegwitNative: null,
-      xpubMagicMultisigSegwitP2sh: null,
-      xpubMagicSegwitNative: null,
-      xpubMagicSegwitP2sh: null,
-      wif: '',
-    },
-    {
-      addressType: 76,
-      addressTypeP2sh: 16,
-      bech32Prefix: 'xc',
-      blockchainLink: null,
-      blocktimeSeconds: 60,
-      cashaddrPrefix: null,
-      coinLabel: 'Stakenet',
-      coinName: 'Stakenet',
-      coinShortcut: 'XSN',
-      curveName: 'secp256k1',
-      decimals: 8,
-      decred: false,
-      defaultFeeB: {
-        economy: 70,
-        high: 200,
-        low: 10,
-        normal: 140,
-      },
-      dustLimit: 1000,
-      extraData: true,
-      forceBip143: false,
-      forkId: null,
-      hashGenesisBlock:
-        '00000c822abdbb23e28f79a49d29b41429737c6c7e15df40d1b1f1b35907ae34',
-      maxAddressLength: 47,
-      maxfeeKb: 11000000000,
-      minAddressLength: 27,
-      minfeeKb: 1000,
-      name: 'Stakenet',
-      segwit: true,
-      shortcut: 'XSN',
-      signedMessageHeader: 'DarkCoin Signed Message:\n',
-      slip44: 199,
-      support: {
-        connect: true,
-        suite: false,
-        trezor1: '1.8.0',
-        trezor2: '2.0.11',
-      },
-      taproot: false,
-      timestamp: false,
-      xprvMagic: 76066276,
-      xpubMagic: 76067358,
-      xpubMagicMultisigSegwitNative: 76067358,
-      xpubMagicMultisigSegwitP2sh: 76067358,
-      xpubMagicSegwitNative: 78792518,
-      xpubMagicSegwitP2sh: 77429938,
-      wif: '',
-    },
-    {
-      addressType: 63,
-      addressTypeP2sh: 5,
-      bech32Prefix: 'sys',
-      blockchainLink: {
-        type: 'blockbook',
-        url: ['https://sys1.bcfn.ca'],
-      },
-      blocktimeSeconds: 60,
-      cashaddrPrefix: null,
-      coinLabel: 'Syscoin',
-      coinName: 'Syscoin',
-      coinShortcut: 'SYS',
-      curveName: 'secp256k1',
-      decimals: 8,
-      decred: false,
-      defaultFeeB: {
-        economy: 700,
-        high: 2000,
-        low: 100,
-        normal: 1400,
-      },
-      dustLimit: 1820,
-      extraData: false,
-      forceBip143: false,
-      forkId: null,
-      hashGenesisBlock:
-        '0000022642db0346b6e01c2a397471f4f12e65d4f4251ec96c1f85367a61a7ab',
-      maxAddressLength: 34,
-      maxfeeKb: 42000000000,
-      minAddressLength: 27,
-      minfeeKb: 10000,
-      name: 'Syscoin',
-      segwit: true,
-      shortcut: 'SYS',
-      signedMessageHeader: 'Syscoin Signed Message:\n',
-      slip44: 57,
-      support: {
-        connect: true,
-        suite: false,
-        trezor1: '1.8.4',
-        trezor2: '2.1.8',
-      },
-      taproot: false,
-      timestamp: false,
-      xprvMagic: 76066276,
-      xpubMagic: 76067358,
-      xpubMagicMultisigSegwitNative: 76067358,
-      xpubMagicMultisigSegwitP2sh: 76067358,
-      xpubMagicSegwitNative: 78792518,
-      xpubMagicSegwitP2sh: 77429938,
-      wif: '',
-    },
-    {
-      addressType: 130,
-      addressTypeP2sh: 30,
-      bech32Prefix: null,
-      blockchainLink: {
-        type: 'blockbook',
-        url: [
-          'https://blockbook.flurbo.xyz',
-          'https://blockbook.unobtanium.uno',
-        ],
-      },
-      blocktimeSeconds: 30,
-      cashaddrPrefix: null,
-      coinLabel: 'Unobtanium',
-      coinName: 'Unobtanium',
-      coinShortcut: 'UNO',
-      curveName: 'secp256k1',
-      decimals: 8,
-      decred: false,
-      defaultFeeB: {
-        economy: 50,
-        high: 160,
-        low: 10,
-        normal: 100,
-      },
-      dustLimit: 546,
-      extraData: false,
-      forceBip143: false,
-      forkId: null,
-      hashGenesisBlock:
-        '000004c2fc5fffb810dccc197d603690099a68305232e552d96ccbe8e2c52b75',
-      maxAddressLength: 34,
-      maxfeeKb: 53000000,
-      minAddressLength: 27,
-      minfeeKb: 1000,
-      name: 'Unobtanium',
-      segwit: false,
-      shortcut: 'UNO',
-      signedMessageHeader: 'Unobtanium Signed Message:\n',
-      slip44: 92,
-      support: {
-        connect: true,
-        suite: false,
-        trezor1: '1.8.4',
-        trezor2: '2.1.6',
-      },
-      taproot: false,
-      timestamp: false,
-      xprvMagic: 76066276,
-      xpubMagic: 76067358,
-      xpubMagicMultisigSegwitNative: null,
-      xpubMagicMultisigSegwitP2sh: null,
-      xpubMagicSegwitNative: null,
-      xpubMagicSegwitP2sh: null,
-      wif: '',
-    },
-    {
-      addressType: 30,
-      addressTypeP2sh: 33,
-      bech32Prefix: null,
-      blockchainLink: null,
-      blocktimeSeconds: 30,
-      cashaddrPrefix: null,
-      coinLabel: 'Verge',
-      coinName: 'Verge',
-      coinShortcut: 'XVG',
-      curveName: 'secp256k1',
-      decimals: 6,
-      decred: false,
-      defaultFeeB: {
-        normal: 10,
-      },
-      dustLimit: 10000,
-      extraData: false,
-      forceBip143: false,
-      forkId: null,
-      hashGenesisBlock:
-        '00000fc63692467faeb20cdb3b53200dc601d75bdfa1001463304cc790d77278',
-      maxAddressLength: 34,
-      maxfeeKb: 550000000000,
-      minAddressLength: 27,
-      minfeeKb: 1000000,
-      name: 'Verge',
-      segwit: false,
-      shortcut: 'XVG',
-      signedMessageHeader: 'Name: Dogecoin Dark\n',
-      slip44: 77,
-      support: {
-        connect: true,
-        suite: false,
-        trezor1: '1.9.3',
-        trezor2: '2.3.3',
-      },
-      taproot: false,
-      timestamp: true,
-      xprvMagic: 35729707,
-      xpubMagic: 36513075,
-      xpubMagicMultisigSegwitNative: null,
-      xpubMagicMultisigSegwitP2sh: null,
-      xpubMagicSegwitNative: null,
-      xpubMagicSegwitP2sh: null,
-      wif: '',
-    },
-    {
-      addressType: 71,
-      addressTypeP2sh: 5,
-      bech32Prefix: 'vtc',
-      blockchainLink: {
-        type: 'blockbook',
-        url: [
-          'https://vtc1.trezor.io',
-          'https://vtc2.trezor.io',
-          'https://vtc3.trezor.io',
-          'https://vtc4.trezor.io',
-          'https://vtc5.trezor.io',
-        ],
-      },
-      blocktimeSeconds: 150,
-      cashaddrPrefix: null,
-      coinLabel: 'Vertcoin',
-      coinName: 'Vertcoin',
-      coinShortcut: 'VTC',
-      curveName: 'secp256k1',
-      decimals: 8,
-      decred: false,
-      defaultFeeB: {
-        normal: 1000,
-      },
-      dustLimit: 54600,
-      extraData: false,
-      forceBip143: false,
-      forkId: null,
-      hashGenesisBlock:
-        '4d96a915f49d40b1e5c2844d1ee2dccb90013a990ccea12c492d22110489f0c4',
-      maxAddressLength: 34,
-      maxfeeKb: 13000000000,
-      minAddressLength: 27,
-      minfeeKb: 100000,
-      name: 'Vertcoin',
-      segwit: true,
-      shortcut: 'VTC',
-      signedMessageHeader: 'Vertcoin Signed Message:\n',
-      slip44: 28,
-      support: {
-        connect: true,
-        suite: true,
-        trezor1: '1.6.1',
-        trezor2: '2.0.5',
-      },
-      taproot: false,
-      timestamp: false,
-      xprvMagic: 76066276,
-      xpubMagic: 76067358,
-      xpubMagicMultisigSegwitNative: 76067358,
-      xpubMagicMultisigSegwitP2sh: 76067358,
-      xpubMagicSegwitNative: 78792518,
-      xpubMagicSegwitP2sh: 77429938,
-      wif: '',
-    },
-    {
-      addressType: 71,
-      addressTypeP2sh: 33,
-      bech32Prefix: 'via',
-      blockchainLink: null,
-      blocktimeSeconds: 24,
-      cashaddrPrefix: null,
-      coinLabel: 'Viacoin',
-      coinName: 'Viacoin',
-      coinShortcut: 'VIA',
-      curveName: 'secp256k1',
-      decimals: 8,
-      decred: false,
-      defaultFeeB: {
-        economy: 7000,
-        high: 20000,
-        low: 1000,
-        normal: 14000,
-      },
-      dustLimit: 54600,
-      extraData: false,
-      forceBip143: false,
-      forkId: null,
-      hashGenesisBlock:
-        '4e9b54001f9976049830128ec0331515eaabe35a70970d79971da1539a400ba1',
-      maxAddressLength: 34,
-      maxfeeKb: 14000000000,
-      minAddressLength: 27,
-      minfeeKb: 1000,
-      name: 'Viacoin',
-      segwit: true,
-      shortcut: 'VIA',
-      signedMessageHeader: 'Viacoin Signed Message:\n',
-      slip44: 14,
-      support: {
-        connect: true,
-        suite: false,
-        trezor1: '1.6.2',
-        trezor2: '2.0.7',
-      },
-      taproot: false,
-      timestamp: false,
-      xprvMagic: 76066276,
-      xpubMagic: 76067358,
-      xpubMagicMultisigSegwitNative: 76067358,
-      xpubMagicMultisigSegwitP2sh: 76067358,
-      xpubMagicSegwitNative: 78792518,
-      xpubMagicSegwitP2sh: 77429938,
-      wif: '',
-    },
-    {
-      addressType: 142,
-      addressTypeP2sh: 145,
-      bech32Prefix: null,
-      blockchainLink: {
-        type: 'blockbook',
-        url: ['https://blockbook.zcore.cash'],
-      },
-      blocktimeSeconds: 120,
-      cashaddrPrefix: null,
-      coinLabel: 'ZCore',
-      coinName: 'ZCore',
-      coinShortcut: 'ZCR',
-      curveName: 'secp256k1',
-      decimals: 8,
-      decred: false,
-      defaultFeeB: {
-        normal: 10,
-      },
-      dustLimit: 546,
-      extraData: false,
-      forceBip143: false,
-      forkId: null,
-      hashGenesisBlock:
-        '695b79c8c234b45b2eeb4722f33373e471c9b686ff78efeb39da95f824a9f81b',
-      maxAddressLength: 95,
-      maxfeeKb: 170000000000,
-      minAddressLength: 34,
-      minfeeKb: 1000,
-      name: 'ZCore',
-      segwit: false,
-      shortcut: 'ZCR',
-      signedMessageHeader: 'DarkNet Signed Message:\n',
-      slip44: 428,
-      support: {
-        connect: true,
-        suite: false,
-        trezor1: '1.8.4',
-        trezor2: '2.1.7',
-      },
-      taproot: false,
-      timestamp: false,
-      xprvMagic: 78791432,
-      xpubMagic: 78792518,
-      xpubMagicMultisigSegwitNative: null,
-      xpubMagicMultisigSegwitP2sh: null,
-      xpubMagicSegwitNative: null,
-      xpubMagicSegwitP2sh: null,
-      wif: '',
-    },
-    {
-      addressType: 7352,
-      addressTypeP2sh: 7357,
-      bech32Prefix: null,
-      blockchainLink: {
-        type: 'blockbook',
-        url: [
-          'https://zec1.trezor.io',
-          'https://zec2.trezor.io',
-          'https://zec3.trezor.io',
-          'https://zec4.trezor.io',
-          'https://zec5.trezor.io',
-        ],
-      },
-      blocktimeSeconds: 150,
-      cashaddrPrefix: null,
-      coinLabel: 'Zcash',
-      coinName: 'Zcash',
-      coinShortcut: 'ZEC',
-      curveName: 'secp256k1',
-      decimals: 8,
-      decred: false,
-      defaultFeeB: {
-        normal: 10,
-      },
-      dustLimit: 546,
-      extraData: true,
-      forceBip143: false,
-      forkId: null,
-      hashGenesisBlock:
-        '00040fe8ec8471911baa1db1266ea15dd06b4a8a5c453883c000b031973dce08',
-      maxAddressLength: 95,
-      maxfeeKb: 51000000,
-      minAddressLength: 35,
-      minfeeKb: 1000,
-      name: 'Zcash',
-      segwit: false,
-      shortcut: 'ZEC',
-      signedMessageHeader: 'Zcash Signed Message:\n',
-      slip44: 133,
-      support: {
-        connect: true,
-        suite: true,
-        trezor1: '1.8.1',
-        trezor2: '2.1.1',
-      },
-      taproot: false,
-      timestamp: false,
-      xprvMagic: 76066276,
-      xpubMagic: 76067358,
-      xpubMagicMultisigSegwitNative: null,
-      xpubMagicMultisigSegwitP2sh: null,
-      xpubMagicSegwitNative: null,
-      xpubMagicSegwitP2sh: null,
-      wif: '',
-    },
-    {
-      addressType: 7461,
-      addressTypeP2sh: 7354,
-      bech32Prefix: null,
-      blockchainLink: null,
-      blocktimeSeconds: 150,
-      cashaddrPrefix: null,
-      coinLabel: 'Zcash Testnet',
-      coinName: 'Zcash Testnet',
-      coinShortcut: 'TAZ',
-      curveName: 'secp256k1',
-      decimals: 8,
-      decred: false,
-      defaultFeeB: {
-        normal: 10,
-      },
-      dustLimit: 546,
-      extraData: true,
-      forceBip143: false,
-      forkId: null,
-      hashGenesisBlock:
-        '05a60a92d99d85997cce3b87616c089f6124d7342af37106edc76126334a2c38',
-      maxAddressLength: 95,
-      maxfeeKb: 10000000,
-      minAddressLength: 35,
-      minfeeKb: 1000,
-      name: 'Zcash Testnet',
-      segwit: false,
-      shortcut: 'TAZ',
-      signedMessageHeader: 'Zcash Signed Message:\n',
-      slip44: 1,
-      support: {
-        connect: true,
-        suite: false,
-        trezor1: '1.8.1',
-        trezor2: '2.1.1',
-      },
-      taproot: false,
-      timestamp: false,
-      xprvMagic: 70615956,
-      xpubMagic: 70617039,
-      xpubMagicMultisigSegwitNative: null,
-      xpubMagicMultisigSegwitP2sh: null,
-      xpubMagicSegwitNative: null,
-      xpubMagicSegwitP2sh: null,
-      wif: '',
-    },
-    {
-      addressType: 61,
-      addressTypeP2sh: 123,
-      bech32Prefix: null,
-      blockchainLink: {
-        type: 'blockbook',
-        url: [
-          'https://blockbook1.bitcoinrh.org',
-          'https://blockbook2.bitcoinrh.org',
-        ],
-      },
-      blocktimeSeconds: 600,
-      cashaddrPrefix: null,
-      coinLabel: 'xRhodium',
-      coinName: 'Brhodium',
-      coinShortcut: 'XRC',
-      curveName: 'secp256k1',
-      decimals: 8,
-      decred: false,
-      defaultFeeB: {
-        economy: 70,
-        high: 200,
-        low: 10,
-        normal: 140,
-      },
-      dustLimit: 546,
-      extraData: false,
-      forceBip143: false,
-      forkId: null,
-      hashGenesisBlock:
-        'baff5bfd9dc43fb672d003ec20fd21428f9282ca46bfa1730d73e1f2c75f5fdd',
-      maxAddressLength: 34,
-      maxfeeKb: 1000000000,
-      minAddressLength: 27,
-      minfeeKb: 1000,
-      name: 'xRhodium',
-      segwit: false,
-      shortcut: 'XRC',
-      signedMessageHeader: 'BitCoin Rhodium Signed Message:\n',
-      slip44: 10291,
-      support: {
-        connect: true,
-        suite: false,
-        trezor1: '1.8.2',
-        trezor2: '2.1.1',
-      },
-      taproot: false,
-      timestamp: false,
-      xprvMagic: 76066276,
-      xpubMagic: 76067358,
-      xpubMagicMultisigSegwitNative: null,
-      xpubMagicMultisigSegwitP2sh: null,
-      xpubMagicSegwitNative: null,
-      xpubMagicSegwitP2sh: null,
-      wif: '',
-    },
-  ],
-  erc20: [],
-  eth: [
-    {
-      blockchainLink: {
-        type: 'blockbook',
-        url: ['https://eth1.trezor.io', 'https://eth2.trezor.io'],
-      },
-      chain: 'eth',
-      chainId: 1,
-      name: 'Ethereum',
-      rskip60: false,
-      shortcut: 'ETH',
-      slip44: 60,
-      support: {
-        connect: true,
-        suite: true,
-        trezor1: '1.6.2',
-        trezor2: '2.0.7',
-      },
-      url: 'https://ethereum.org',
-    },
-    {
-      blockchainLink: null,
-      chain: 'exp',
-      chainId: 2,
-      name: 'Expanse Network',
-      rskip60: false,
-      shortcut: 'EXP',
-      slip44: 40,
-      support: {
-        connect: true,
-        suite: false,
-        trezor1: '1.6.2',
-        trezor2: '2.0.7',
-      },
-      url: 'https://expanse.tech',
-    },
-    {
-      blockchainLink: {
-        type: 'blockbook',
-        url: ['https://ropsten1.trezor.io', 'https://ropsten2.trezor.io'],
-      },
-      chain: 'rop',
-      chainId: 3,
-      name: 'Ethereum Testnet Ropsten',
-      rskip60: false,
-      shortcut: 'tROP',
-      slip44: 1,
-      support: {
-        connect: true,
-        suite: true,
-        trezor1: '1.6.2',
-        trezor2: '2.0.7',
-      },
-      url: 'https://github.com/ethereum/ropsten',
-    },
-    {
-      blockchainLink: null,
-      chain: 'rin',
-      chainId: 4,
-      name: 'Ethereum Testnet Rinkeby',
-      rskip60: false,
-      shortcut: 'tRIN',
-      slip44: 1,
-      support: {
-        connect: true,
-        suite: false,
-        trezor1: '1.6.2',
-        trezor2: '2.0.7',
-      },
-      url: 'https://www.rinkeby.io',
-    },
-    {
-      blockchainLink: null,
-      chain: 'ubq',
-      chainId: 8,
-      name: 'Ubiq',
-      rskip60: false,
-      shortcut: 'UBQ',
-      slip44: 108,
-      support: {
-        connect: true,
-        suite: false,
-        trezor1: '1.6.2',
-        trezor2: '2.0.7',
-      },
-      url: 'https://ubiqsmart.com',
-    },
-    {
-      blockchainLink: null,
-      chain: 'rsk',
-      chainId: 30,
-      name: 'RSK',
-      rskip60: true,
-      shortcut: 'RBTC',
-      slip44: 137,
-      support: {
-        connect: true,
-        suite: false,
-        trezor1: '1.6.2',
-        trezor2: '2.0.7',
-      },
-      url: 'https://rsk.co',
-    },
-    {
-      blockchainLink: null,
-      chain: 'trsk',
-      chainId: 31,
-      name: 'RSK Testnet',
-      rskip60: false,
-      shortcut: 'tRBTC',
-      slip44: 1,
-      support: {
-        connect: true,
-        suite: false,
-        trezor1: '1.6.2',
-        trezor2: '2.0.7',
-      },
-      url: 'https://rsk.co',
-    },
-    {
-      blockchainLink: null,
-      chain: 'kov',
-      chainId: 42,
-      name: 'Ethereum Testnet Kovan',
-      rskip60: false,
-      shortcut: 'tKOV',
-      slip44: 1,
-      support: {
-        connect: true,
-        suite: false,
-        trezor1: '1.6.2',
-        trezor2: '2.0.7',
-      },
-      url: 'https://kovan-testnet.github.io/website',
-    },
-    {
-      blockchainLink: null,
-      chain: 'go',
-      chainId: 60,
-      name: 'GoChain',
-      rskip60: false,
-      shortcut: 'GO',
-      slip44: 6060,
-      support: {
-        connect: true,
-        suite: false,
-        trezor1: '1.6.2',
-        trezor2: '2.0.7',
-      },
-      url: 'https://gochain.io',
-    },
-    {
-      blockchainLink: {
-        type: 'blockbook',
-        url: ['https://etc1.trezor.io', 'https://etc2.trezor.io'],
-      },
-      chain: 'etc',
-      chainId: 61,
-      name: 'Ethereum Classic',
-      rskip60: false,
-      shortcut: 'ETC',
-      slip44: 61,
-      support: {
-        connect: true,
-        suite: true,
-        trezor1: '1.6.2',
-        trezor2: '2.0.7',
-      },
-      url: 'https://ethereumclassic.org',
-    },
-    {
-      blockchainLink: null,
-      chain: 'ella',
-      chainId: 64,
-      name: 'Ellaism',
-      rskip60: false,
-      shortcut: 'ELLA',
-      slip44: 163,
-      support: {
-        connect: true,
-        suite: false,
-        trezor1: '1.6.2',
-        trezor2: '2.0.7',
-      },
-      url: 'https://ellaism.org',
-    },
-    {
-      blockchainLink: null,
-      chain: 'mix',
-      chainId: 76,
-      name: 'Mix',
-      rskip60: false,
-      shortcut: 'MIX',
-      slip44: 76,
-      support: {
-        connect: true,
-        suite: false,
-        trezor1: '1.7.2',
-        trezor2: '2.0.10',
-      },
-      url: 'https://mix-blockchain.org',
-    },
-    {
-      blockchainLink: null,
-      chain: 'clo',
-      chainId: 820,
-      name: 'Callisto',
-      rskip60: false,
-      shortcut: 'CLO',
-      slip44: 820,
-      support: {
-        connect: true,
-        suite: false,
-        trezor1: '1.6.2',
-        trezor2: '2.0.7',
-      },
-      url: 'https://callisto.network',
-    },
-    {
-      blockchainLink: null,
-      chain: 'egem',
-      chainId: 1987,
-      name: 'EtherGem',
-      rskip60: false,
-      shortcut: 'EGEM',
-      slip44: 1987,
-      support: {
-        connect: true,
-        suite: false,
-        trezor1: '1.6.2',
-        trezor2: '2.0.7',
-      },
-      url: 'https://egem.io',
-    },
-    {
-      blockchainLink: null,
-      chain: 'esn',
-      chainId: 31102,
-      name: 'Ethersocial Network',
-      rskip60: false,
-      shortcut: 'ESN',
-      slip44: 31102,
-      support: {
-        connect: true,
-        suite: false,
-        trezor1: '1.6.3',
-        trezor2: '2.0.8',
-      },
-      url: 'https://ethersocial.org',
-    },
-    {
-      blockchainLink: null,
-      chain: 'aka',
-      chainId: 200625,
-      name: 'Akroma',
-      rskip60: false,
-      shortcut: 'AKA',
-      slip44: 200625,
-      support: {
-        connect: true,
-        suite: false,
-        trezor1: '1.6.3',
-        trezor2: '2.0.8',
-      },
-      url: 'https://akroma.io',
-    },
-    {
-      blockchainLink: null,
-      chain: 'etho',
-      chainId: 1313114,
-      name: 'Etho Protocol',
-      rskip60: false,
-      shortcut: 'ETHO',
-      slip44: 1313114,
-      support: {
-        connect: true,
-        suite: false,
-        trezor1: '1.6.3',
-        trezor2: '2.0.8',
-      },
-      url: 'https://ethoprotocol.com',
-    },
-    {
-      blockchainLink: null,
-      chain: 'music',
-      chainId: 7762959,
-      name: 'Musicoin',
-      rskip60: false,
-      shortcut: 'MUSIC',
-      slip44: 184,
-      support: {
-        connect: true,
-        suite: false,
-        trezor1: '1.6.3',
-        trezor2: '2.0.8',
-      },
-      url: 'https://musicoin.tw',
-    },
-    {
-      blockchainLink: null,
-      chain: 'pirl',
-      chainId: 3125659152,
-      name: 'Pirl',
-      rskip60: false,
-      shortcut: 'PIRL',
-      slip44: 164,
-      support: {
-        connect: true,
-        suite: false,
-        trezor1: '1.6.3',
-        trezor2: '2.0.8',
-      },
-      url: 'https://pirl.io',
-    },
-  ],
-  misc: [
-    {
-      blockchainLink: {
-        type: 'blockfrost',
-        url: ['wss://trezor-cardano-mainnet.blockfrost.io'],
-      },
-      curve: 'ed25519',
-      decimals: 6,
-      name: 'Cardano',
-      shortcut: 'ADA',
-      slip44: 1815,
-      support: {
-        connect: true,
-        suite: false,
-        trezor1: false,
-        trezor2: '2.0.8',
-      },
-    },
-    {
-      blockchainLink: null,
-      curve: 'secp256k1',
-      decimals: 8,
-      name: 'Binance Chain',
-      shortcut: 'BNB',
-      slip44: 714,
-      support: {
-        connect: true,
-        suite: false,
-        trezor1: false,
-        trezor2: '2.1.5',
-      },
-    },
-    {
-      blockchainLink: null,
-      curve: 'secp256k1',
-      decimals: 4,
-      name: 'EOS',
-      shortcut: 'EOS',
-      slip44: 194,
-      support: {
-        connect: true,
-        suite: false,
-        trezor1: false,
-        trezor2: '2.1.1',
-      },
-    },
-    {
-      blockchainLink: {
-        type: 'blockfrost',
-        url: ['wss://trezor-cardano-testnet.blockfrost.io'],
-      },
-      curve: 'ed25519',
-      decimals: 6,
-      name: 'Cardano Testnet',
-      shortcut: 'tADA',
-      slip44: 1815,
-      support: {
-        connect: true,
-        suite: false,
-        trezor1: false,
-        trezor2: '2.0.8',
-      },
-    },
-    {
-      blockchainLink: {
-        type: 'ripple',
-        url: ['wss://s.altnet.rippletest.net'],
-      },
-      curve: 'secp256k1',
-      decimals: 6,
-      name: 'Ripple Testnet',
-      shortcut: 'tXRP',
-      slip44: 1,
-      support: {
-        connect: true,
-        suite: true,
-        trezor1: false,
-        trezor2: '2.0.8',
-      },
-    },
-    {
-      blockchainLink: null,
-      curve: 'ed25519',
-      decimals: 7,
-      name: 'Stellar',
-      shortcut: 'XLM',
-      slip44: 148,
-      support: {
-        connect: true,
-        suite: false,
-        trezor1: '1.7.1',
-        trezor2: '2.0.8',
-      },
-    },
-    {
-      blockchainLink: {
-        type: 'ripple',
-        url: [
-          'wss://s1.ripple.com',
-          'wss://s-east.ripple.com',
-          'wss://s-west.ripple.com',
-        ],
-      },
-      curve: 'secp256k1',
-      decimals: 6,
-      name: 'Ripple',
-      shortcut: 'XRP',
-      slip44: 144,
-      support: {
-        connect: true,
-        suite: true,
-        trezor1: false,
-        trezor2: '2.0.8',
-      },
-    },
-    {
-      blockchainLink: null,
-      curve: 'ed25519',
-      decimals: 6,
-      name: 'Tezos',
-      shortcut: 'XTZ',
-      slip44: 1729,
-      support: {
-        connect: true,
-        suite: false,
-        trezor1: false,
-        trezor2: '2.0.8',
-      },
-    },
-  ],
-};
